### 2021-01-08 #420 @temospena
slopes
Submitting Author: Rosa Flx (@temospena)  and Robin Lovelace (@robinlovelace)
Repository:  https://github.com/ITSLeeds/slopes
Version submitted: 0.0.1
Version accepted: TBD

---

```
Package: slopes
Title: Calculate Slopes of Roads, Rivers and Trajectories
Version: 0.0.0.9000
    person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut","cre"),
    person("Joey", "Talbot", email = "j.d.talbot@leeds.ac.uk", role = c("aut"),
    person("Rosa", "Félix", email = "rosamfelix@tecnico.ulisboa.pt", role = c("aut"),
    )
Description: Functions and example data to support research into
  'longitudinal gradient' (slope) of linear geographic entities, as defined in the context of rivers
  by Cohen et al. (2018) <doi:10.1016/j.jhydrol.2018.06.066>.
  The package was initially developed to research road slopes but has also been
  used to calculate and visualise slopes of rivers and, using an open dataset
  published by Javier Ariza-López et al. (2019) <doi:10.1038/s41597-019-0147-x>, trajectories representing
  movement on roads.
  The package takes two main types of input data for slope calculation: vector geographic
  objects representing linear features, and raster geographic objects with elevation values
  (which can be downloaded using functionality in the package)
  representing a continuous terrain surface.
    sf,
    raster,
    methods
    geodist,
    pbapply,
    terra,
    colorspace,
    ceramic,
    bookdown

```

The `slopes` package retreives elevation data via an interface to the `ceramic` package, enabling estimation of hilliness for routes anywhere worldwide even when local elevation data is lacking.

The main category is geospatial data: the package takes geographic lines objects and returns elevation data per vertex (providing the output as a 3D point geometry in the `sf` package by default) and per line feature (providing average gradient by default).

The target audience is academic researchers, practitioners, consultants and citizens interested in elevation profiles and slopes worldwide.
A growing number of people working with geospatial data and require accurate estimates of gradient.
An example of the demand for data provided by the package is a map showing gradients across Sao Paulo (Brazil, see image below) that has received more the 100 'likes' on Twitter and generated convervations: https://twitter.com/DanielGuth/status/1347270685161304069

**Transport planning practitioners** require accurate estimates of roadway gradient for estimating energy consumption, safety and mode shift potential in hilly cities (such as Lisbon, the case study city used in the README and examples in the documentation).
**Natural hazard researchers and risk assessors** require estimates of linear gradient to inform safety and mitigation plans associated with project on hilly terrain.
We believe that **aquatic ecologists**, flooding researchers and others could benefit from estimates of river gradient to support modelling of storm hydrographs, although we do not have experience in these domains.

The package fills a niche in R's geospatial package ecosystem. It is the only R package for easy slope computation of a line segment (or several) .
It is also the only open source tool dedicated to the estimation of slopes on linear features in any language, as far as we are aware.
We have benchmarked the results against a market leading proprietary implementation (ArcMap's 3D Analyst Extension).
We would like to add more cross-comparisons of slope estimates obtained with this package and other implementations.

To add - see https://github.com/ropensci/software-review/issues/395
**VERIFY**:_We believe the package complies with ethical guidelines in the Internet Research: [Ethical Guidelines 3.0 document](https://aoir.org/reports/ethics3.pdf).
It makes it easier for researchers to access and make use of data that is already in the public domain, under the conditions of the adhering to the conditions of the [OdBL](https://opendatacommons.org/licenses/odbl/)._

(We have a test suite but with low coverage, we aim to increase coverage of functions tested.)
