### 2021-01-08 ITSLeeds/slopes#21 @Robinlovelace
Function to get elevation associated with points
As an alternative to the approach of downloading the DEM for a whole area. May also need subsampling of vertices.
