### 2021-01-11 95a9201
Update terra extract code, fixing #22
