+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   slopes: slopes.html
   verification: verification.html
-last_built: 2020-11-16T15:00Z
+last_built: 2021-01-11T00:05Z
 
