### 2021-01-11 93197f7
Do not print TRUE when calling has_terra()
