### 2021-01-14 #420 @annakrystalli
### Editor checks:

---

#### Editor comments

Many thanks again for the submission of this useful package @temospena & @Robinlovelace. I've now completed initial editors checks. Comments below, the most pressing one regarding test coverage (see below).

### `devtools::check(pkg_dir)`

Checks throw up a single note. Seems minor but would need addressing seeing as you wish to submit to CRAN.
```
checking Rd cross-references ... NOTE
  Package unavailable to check Rd xrefs: ‘stplanr’

0 errors ✓ | 0 warnings ✓ | 1 note x
```

### `goodpractice::gp()`

A few  points to look into were returned by `goodpractice::gp()`. The most critical being the **low test coverage**. To accept a package we require a minimum of 75% coverage. Is there a reason only 10% of code lines are currently tested?

```
  ✖ write unit tests for all functions, and all
    package code in general. 10% of code lines are covered by
    test cases.

    R/plot_slope.R:30:NA
    R/plot_slope.R:31:NA
    R/plot_slope.R:32:NA
    R/plot_slope.R:33:NA
    R/plot_slope.R:34:NA
    ... and 130 more lines

  ✖ add a "BugReports" field to DESCRIPTION, and
    point it to a bug tracker. Many online code hosting
    services provide bug trackers for free,
    https://github.com, https://gitlab.com, etc.
  ✖ use '<-' for assignment instead of '='. '<-' is
    the standard, and R users and developers are used it and
    it is easier to read your code for them if you use '<-'.

    R/plot_slope.R:13:12
    R/plot_slope.R:30:5
    R/plot_slope.R:31:5
    R/plot_slope.R:32:5
    R/plot_slope.R:33:5
    ... and 96 more lines

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor windows that
    are about 80 characters wide. Try make your lines shorter
    than 80 characters

    R/data.R:3:1
    R/data.R:5:1
    R/data.R:15:1
    R/data.R:18:1
    R/data.R:29:1
    ... and 24 more lines

  ✖ fix this R CMD check NOTE: Namespace in Imports
    field not imported from: ‘methods’ All declared Imports
    should be used.
  ✖ avoid 'T' and 'F', as they are just variables
    which are set to the logicals 'TRUE' and 'FALSE' by
    default, but are not reserved words and hence can be
    overwritten by the user.  Hence, one should always use
    'TRUE' and 'FALSE' for the logicals.

    R/z.R:NA:NA
```

More detailed breakdown of the low coverage revealled by`covr::package_coverage()`
```
slopes Coverage: 10.60%
R/plot_slope.R: 0.00%
R/slope_get.R: 0.00%
R/z.R: 16.67%
R/slopes.R: 16.90%
```
i've also highlighted a few potential spelling issues returned by `devtools::spell_check()`

```
datast            README.md:87
                  README.Rmd:83
formular          slopes.Rmd:96
implemenation     slopes.Rmd:85
introdution       slopes.Rmd:2
lenght            slopes.Rmd:51
purpouse          slopes.Rmd:33
```
