### 2021-01-22 03be027
Add validation code starter
