### 2021-01-22 c6063d1
Update magnolia data
