### 2021-01-25 48a4e86
Add starter for #23
