### 2021-01-25 ea8ec11
Add magnolia_xyz dataset for #23
