+++ man/slope_xyz.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/slopes.R
+\name{slope_xyz}
+\alias{slope_xyz}
+\title{Extract slopes from xyz dataframe or sf objects}
+\usage{
+slope_xyz(r_xyz)
+}
+\description{
+Extract slopes from xyz dataframe or sf objects
+}
