+++ NAMESPACE
@@ -1,9 +1,13 @@
 # Generated by roxygen2: do not edit by hand
 
+export("=")
+export("]")
+export("slopes::lisbon_route_3d[1:3,")
 export(elevation_extract)
 export(elevations_get)
 export(plot_dz)
 export(plot_slope)
+export(r_xyz)
 export(sequential_dist)
 export(slope_3d)
 export(slope_distance)
