### 2021-01-25 71c94e1
slope_xyyz works
