+++ NAMESPACE
@@ -1,13 +1,9 @@
 # Generated by roxygen2: do not edit by hand
 
-export("=")
-export("]")
-export("slopes::lisbon_route_3d[1:3,")
 export(elevation_extract)
 export(elevations_get)
 export(plot_dz)
 export(plot_slope)
-export(r_xyz)
 export(sequential_dist)
 export(slope_3d)
 export(slope_distance)
@@ -17,6 +13,7 @@ export(slope_matrix)
 export(slope_matrix_weighted)
 export(slope_raster)
 export(slope_vector)
+export(slope_xyz)
 export(slope_z_end)
 export(slope_z_max)
 export(slope_z_mean)
