### 2021-01-26 7b4a8b2
comparing slopes
test with seattle data
