### 2021-01-26 ba310c2
Fix lonlat issue with slope_raster
