+++ man/magnolia_xyz.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data.R
-\docType{data}
-\name{magnolia_xyz}
-\alias{magnolia_xyz}
-\title{Road segments in Magnolia, Seattle}
-\format{
-An object of class \code{sf}
-}
-\source{
-Produced by the website \href{https://data-seattlecitygis.opendata.arcgis.com/datasets/seattle-streets}{data-seattlecitygis.opendata.arcgis.com}
-}
-\usage{
-magnolia_xyz
-}
-\description{
-A dataset representing road segments in the Magnolia area of Seattle
-with X, Y and Z (elevation) dimensions for each coordinate.
-}
-\examples{
-names(magnolia_xyz)
-plot(magnolia_xyz["SLOPE_PCT"])
-}
-\keyword{datasets}
