### 2021-01-26 3710b79
original vs. ceramic vs. SRTM
test seattle: correlations
