+++ README.Rmd
@@ -51,13 +51,6 @@ We will also load the `sf` library:
 library(sf)
 ```
 
-
-<!-- We will also use the `sf` package for representing road segments: -->
-
-<!-- ```{r} -->
-<!-- library(sf) -->
-<!-- ``` -->
-
 The minimum data requirements for using the package are elevation points, either as a vector, a matrix or as a digital elevation model (DEM) encoded as a raster dataset.
 Typically you will also have a geographic object representing the roads or similar features.
 These two types of input data are represented in the code output and plot below.
@@ -83,7 +76,7 @@ summary(lisbon_road_segments$slope)
 This created a new column, `slope` that represents the average, distance weighted slope associated with each road segment.
 The units represent the percentage incline, that is the change in elevation divided by distance.
 The summary of the result tells us that the average gradient of slopes in the example data is just over 5%.
-This result is equivalent to that returned by ESRI's `Slope_3d()` in the [3D Analyst extension](https://desktop.arcgis.com/en/arcmap/10.3/tools/3d-analyst-toolbox/slope.htm), with a correlation between the ArcMap implementation and our implementation of more than 0.95 on our test datast (we find higher correlations on larger datasets):
+This result is equivalent to that returned by ESRI's `Slope_3d()` in the [3D Analyst extension](https://desktop.arcgis.com/en/arcmap/10.3/tools/3d-analyst-toolbox/slope.htm), with a correlation between the ArcMap implementation and our implementation of more than 0.95 on our test dataset (we find higher correlations on larger datasets):
 
 ```{r}
 cor(
