+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   slopes: slopes.html
   verification: verification.html
-last_built: 2021-01-29T19:52Z
+last_built: 2021-02-02T18:18Z
 
