### 2021-02-02 4e2ebc5
spell cheking slopes.Rmd vignette
