### 2021-02-02 e0c6715
Update slope_get function, fix #25
See https://github.com/hypertidy/geodist/issues/33
