+++ tests/testthat/test-z.R
@@ -3,6 +3,11 @@ test_that("Functions on Z values work", {
   expect_equal(class(slope_z_value(x)), "numeric")
   x = slopes::lisbon_route
   expect_error(slope_z_value(x))
+  expect_error(slope_z_start(x))
+  expect_error(slope_z_end(x))
+  expect_error(slope_z_mean(x))
+  expect_error(slope_z_max(x))
+  expect_error(slope_z_min(x))
   # x_slope =
   # expect_is()
 })
