+++ tests/testthat/test-get.R
@@ -1,6 +1,8 @@
-test_that("slope_* functions work", {
+test_that("functions to get elevations from sf objects work", {
+  if(nchar(Sys.getenv("MAPBOX")) < 8)
+    skip(message = "Skipping test, MAPBOX token in .Renviron needed")
   r = cyclestreets_route
   e = elevations_get(r)
-  skip_if(nchar(Sys.getenv("MAPBOX") < 8))
   expect_equal(class(e)[1], "RasterLayer")
 })
+
