### 2021-02-03 61bbda4
Add test for plots
