### 2021-02-03 c949930
Refactor plot_dz()
