### 2021-02-03 83cd94f
Refactor plots - for coverage
