### 2021-02-03 147f09c
More tests for plot slope
