### 2021-02-03 c4df8f9
Fix issues flagged by goodpractice
