+++ R/data.R
@@ -1,8 +1,10 @@
 #' Elevation in central Lisbon, Portugal
 #'
-#' A dataset containing elevation in and around Lisbon with a geographic resolution of 10m.
+#' A dataset containing elevation in and around Lisbon
+#' with a geographic resolution of 10m.
 #'
-#' @format A raster dataset containing elevation above sea level in a 1km bounding box in Lisbon, Portugal.
+#' @format A raster dataset containing elevation above sea level
+#'   in a 1km bounding box in Lisbon, Portugal.
 #' @source \url{https://github.com/rspatial/terra/issues/29}
 #' @examples
 #' names(dem_lisbon_raster)
@@ -12,10 +14,12 @@
 
 #' Road segments in Lisbon
 #'
-#' A dataset representing road segments in Lisbon, with X, Y and Z (elevation) dimensions for each coordinate.
+#' A dataset representing road segments in Lisbon,
+#'   with X, Y and Z (elevation) dimensions for each coordinate.
 #'
 #' @format An object of class `sf`
-#' @source Produced by ESRI's [3D Analyst extension](https://pro.arcgis.com/en/pro-app/help/analysis/3d-analyst/get-started-with-3d-analyst-in-pro.htm)
+#' @source Produced by ESRI's
+#'  [3D Analyst extension](https://pro.arcgis.com/en/pro-app/help/analysis/)
 #' @examples
 #' names(lisbon_road_segments)
 #' plot(lisbon_road_segments["Avg_Slope"])
@@ -26,7 +30,8 @@
 #' A single road segment and a 3d version.
 #'
 #' @format An object of class `sf`
-#' @source Produced by ESRI's [3D Analyst extension](https://pro.arcgis.com/en/pro-app/help/analysis/3d-analyst/get-started-with-3d-analyst-in-pro.htm)
+#' @source Produced by ESRI's
+#'  [3D Analyst extension](https://pro.arcgis.com/en/pro-app/help/analysis/)
 #' @aliases lisbon_road_segment_3d lisbon_route lisbon_route_3d
 #' @examples
 #' lisbon_road_segment
@@ -35,7 +40,8 @@
 
 #' A journey from CycleStreets.net
 #'
-#' Road segments representing suggested route to cycle in Leeds, UK.
+#' Road segments representing suggested route to cycle
+#' in Leeds, UK.
 #'
 #' @format An object of class `sf`
 #' @source CycleStreets.net
@@ -55,7 +61,8 @@
 #' with X, Y and Z (elevation) dimensions for each coordinate.
 #'
 #' @format An object of class `sf`
-#' @source Produced by the website [data-seattlecitygis.opendata.arcgis.com](https://data-seattlecitygis.opendata.arcgis.com/datasets/seattle-streets)
+#' @source Produced by the website
+#'   [data-seattlecitygis.opendata.arcgis.com](https://data-seattlecitygis.opendata.arcgis.com/datasets/seattle-streets)
 #' @examples
 #' names(magnolia_xy)
 #' plot(magnolia_xy["SLOPE_PCT"])
