### 2021-02-03 f9d2454
More plot code refactoring
