+++ tests/testthat/test-slopes.R
@@ -51,6 +51,8 @@ test_that("slope_* functions work", {
 
   r = lisbon_road_segments[204, ]
   r3d = slope_3d(r, e)
-  expect_equal(sf::st_z_range(r3d$geom), c(86, 92), ignore_attr = TRUE, tolerance = 1)
+  expect_equal(sf::st_z_range(r3d$geom), c(86, 92), ignore_attr = TRUE, tolerance = 10)
+  r3d2 = slope_3d(r)
+  expect_equal(sf::st_z_range(r3d2$geom), c(86, 92), ignore_attr = TRUE, tolerance = 10)
 
 })
