### 2021-02-03 fb3ace6
Add another plot test
