+++ DESCRIPTION
@@ -22,6 +22,7 @@ Description: Functions and example data to support research into
   representing a continuous terrain surface.
 License: GPL-3
 URL: https://github.com/itsleeds/slopes/, https://itsleeds.github.io/slopes/
+BugReports: https://github.com/itsleeds/slopes/issues
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
