### 2021-02-03 b33e889
Skip test that requires mapbox key
