### 2021-02-03 15f033e
Remove link to stplanr for checks
