### 2021-02-08 #420 @temospena
Hi @annakrystalli
The package has coverage of **75%** when tests run on a computer that has a MAPBOX api key in the environment (see: https://github.com/ITSLeeds/slopes/issues/26 ) - we would like to add this to the tests but we are not sure how... any clues?

As it is (**62%**), you may see that `slope_get` is the one with zero coverage, when without an api key.
![image](https://user-images.githubusercontent.com/39107166/107249467-4e04af80-6a2b-11eb-9064-3909ed9dd9db.png)
https://codecov.io/gh/itsleeds/slopes/tree/master/R

Thanks!
