+++ man/magnolia_xy.Rd
@@ -8,8 +8,9 @@
 An object of class \code{sf}
 }
 \source{
-Produced by the website
-\href{https://data-seattlecitygis.opendata.arcgis.com/datasets/seattle-streets}{data-seattlecitygis.opendata.arcgis.com}
+Accessed in early 2021 from the \code{seattle-streets} layer from the
+\href{https://data-seattlecitygis.opendata.arcgis.com/}{data-seattlecitygis}
+website.
 }
 \usage{
 magnolia_xy
