+++ man/elevation_extract.Rd
@@ -33,7 +33,8 @@ e = dem_lisbon_raster
 elevation_extract(m, e)
 elevation_extract(m, e, method = "simple")
 # uncomment the following lines to test with terra (experimental)
-# u = "https://github.com/ITSLeeds/slopes/releases/download/0.0.0/dem_lisbon.tif"
+# u = paste0("https://github.com/ITSLeeds/slopes/",
+#    "releases/download/0.0.0/dem_lisbon.tif" )
 # if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
 # et = terra::rast("dem_lisbon.tif")
 # elevation_extract(m, et)
