### 2021-02-08 #420 @Robinlovelace
Latest `gp()` results below. One question, any idea how to fix the one about 'methods' being used? `methods::*()` functions are used in the package so a bit :confused:  about that one. Other than that, it's definitely an improvement and can get even better no doubt in terms of the tests on computers that don't have the MAPBOX API key environment variable set.

```
── GP slopes ──────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 75% of code lines are covered by test cases.

    R/plot_slope.R:30:NA
    R/plot_slope.R:31:NA
    R/plot_slope.R:77:NA
    R/plot_slope.R:78:NA
    R/plot_slope.R:79:NA
    ... and 34 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the standard,
    and R users and developers are used it and it is easier to
    read your code for them if you use '<-'.

    R/plot_slope.R:13:12
    R/plot_slope.R:30:6
    R/plot_slope.R:59:9
    R/plot_slope.R:79:7
    R/plot_slope.R:80:9
    ... and 136 more lines

  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘methods’ All declared Imports should be used.
  ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF
    version. This typically indicates Rd problems.
```
