### 2021-02-12 #420 @annakrystalli
Hello @temospena & @Robinlovelace, thanks for all the great work done so far on the tests! I'll try and answer your questions:

- **mapbox API key**. I think the simplest way would be ass the API key to your repo's [.secrets](https://docs.github.com/en/actions/reference/encrypted-secrets)?
- **testing plots** I think it's fine at this point to proceed without tests for the time being and have that as an enhancement. Perhaps a simple tests you could include would be to test against a reference snapshot of a saved plot with [`expect_snapshot_file()`](https://testthat.r-lib.org/reference/expect_snapshot_file.html)
- `methods` dependency: It looks like `methods` is a [special case](https://github.com/hadley/r-pkgs/issues/203). Having a look at this [more detailed documentation on namespacing](https://r-pkgs.org/namespace.html) and in particular the [S4 section](https://r-pkgs.org/namespace.html#import-s4) it looks like the workaround is to import the `methods` functions you need in the function roxygen2 documentation.  I've opened a PR that seems to fix this issue.

- On that subject of dependencies, why is `ceramic` not specified as an import? It's being within a function with a note to install it manually if missing? If there is some sort of issue you are trying to get round, it's best to outline any necessary additional installation instructions in the installation part of the `README`.

### Test failure on `check()`

I'm getting the following failed test now:

```r
── Failure (test-plots.R:24:3): plotting functions work ────────────────────────
make_pal(p, b[1:4]) (`actual`) not equal to c("#FF0000", "#00FF00", "#0000FF") (`expected`).

`actual`:   "#FF0000FF" "#00FF00FF" "#0000FFFF"
`expected`: "#FF0000"   "#00FF00"   "#0000FF"
```
