+++ .github/workflows/R-CMD-check.yaml
@@ -30,6 +30,7 @@ jobs:
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       RSPM: ${{ matrix.config.rspm }}
+      MAPBOX: ${{ secrets.MAPBOX }}
 
     steps:
       - uses: actions/checkout@v2
