+++ .github/workflows/test-coverage.yaml
@@ -15,6 +15,8 @@ jobs:
     runs-on: macOS-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      MAPBOX: ${{ secrets.MAPBOX }}
+
     steps:
       - uses: actions/checkout@v2
 
