### 2021-03-18 #420 @Robinlovelace
Good news on this
>     * **mapbox API key**. I think the simplest way would be add the API key to your repo's [.secrets](https://docs.github.com/en/actions/reference/encrypted-secrets)?

Fixed as documented in the latest links in https://github.com/ITSLeeds/slopes/issues/18 or at least it will be when the `sf` package works on Mac, which should be in the next day or so when the binaries update:

![image](https://user-images.githubusercontent.com/1825120/111706966-23ccbb80-883b-11eb-8906-20052001a187.png)

>     * **testing plots** I think it's fine at this point to proceed without tests for the time being and have that as a potential future enhancement. Perhaps a simple tests you could include would be to test against a reference snapshot of a saved plot with [`expect_snapshot_file()`](https://testthat.r-lib.org/reference/expect_snapshot_file.html)

Happy to proceed at this point :+1: took a look at that and it seemed hard + plots may change in near future

>     * `methods` dependency: It looks like `methods` is a [special case](https://github.com/hadley/r-pkgs/issues/203). Having a look at this [more detailed documentation on namespacing](https://r-pkgs.org/namespace.html) and in particular the [S4 section](https://r-pkgs.org/namespace.html#import-s4) it looks like the workaround is to import the `methods` functions you need in the function roxygen2 documentation.  I've opened a PR that seems to fix this issue.

Thank you :pray: https://github.com/ITSLeeds/slopes/pull/27

>     * On that subject of dependencies, why is `ceramic` not specified as an import? It's being used within a function with a message to install it manually if missing? If there is some sort of issue you are trying to get round, it's best to outline any necessary additional installation instructions in the installation part of the `README`.

`ceramic` does not have core functionality and is a bit fiddly to set-up. Wanting to minimise dependencies for fast user install times. Could promote from `Suggests` to `Imports` if the +s of doing so outweigh the fairly minor -s of extra install time and dependencies for some users.

> I'm getting the following failed test...

Cannot reproduce that, it passes on my set-up as documented below.

```
> testthat::test_file("tests/testthat/test-plots.R")

══ Testing test-plots.R ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
[ FAIL 0 | WARN 0 | SKIP 0 | PASS 7 ] Done!
> sessionInfo()
R version 4.0.4 (2021-02-15)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 20.04.2 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/blas/libblas.so.3.9.0
LAPACK: /usr/lib/x86_64-linux-gnu/lapack/liblapack.so.3.9.0

locale:
 [1] LC_CTYPE=en_GB.UTF-8       LC_NUMERIC=C               LC_TIME=en_GB.UTF-8        LC_COLLATE=en_GB.UTF-8     LC_MONETARY=en_GB.UTF-8
 [6] LC_MESSAGES=en_GB.UTF-8    LC_PAPER=en_GB.UTF-8       LC_NAME=C                  LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_GB.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] testthat_3.0.2    slopes_0.0.0.9000

 [1] Rcpp_1.0.6         lattice_0.20-41    prettyunits_1.1.1  class_7.3-18       ps_1.6.0           assertthat_0.2.1   rprojroot_2.0.2
 [8] digest_0.6.27      utf8_1.2.1         R6_2.5.0           evaluate_0.14      e1071_1.7-6        pillar_1.5.1       rlang_0.4.10
[15] rstudioapi_0.13    callr_3.5.1        raster_3.4-5       rmarkdown_2.7      desc_1.3.0         devtools_2.3.2     stringr_1.4.0
[22] proxy_0.4-25       compiler_4.0.4     xfun_0.22          pkgconfig_2.0.3    pkgbuild_1.2.0     htmltools_0.5.1.1  tidyselect_1.1.0
[29] tibble_3.1.0       roxygen2_7.1.1     codetools_0.2-18   fansi_0.4.2        crayon_1.4.1       dplyr_1.0.5        withr_2.4.1
[36] sf_0.9-8           waldo_0.2.5        commonmark_1.7     grid_4.0.4         lifecycle_1.0.0    DBI_1.1.1          magrittr_2.0.1
[43] units_0.7-1        KernSmooth_2.23-18 cli_2.3.1          stringi_1.5.3      cachem_1.0.4       diffobj_0.3.3      fs_1.5.0
[50] remotes_2.2.0      sp_1.4-5           xml2_1.3.2         ellipsis_0.3.1     generics_0.1.0     vctrs_0.3.6.9000   rematch2_2.1.2
[57] tools_4.0.4        rcmdcheck_1.3.3    glue_1.4.2         purrr_0.3.4        yaml_2.2.1         processx_3.4.5     pkgload_1.2.0
[64] fastmap_1.1.0      colorspace_2.0-0   xopen_1.0.0        sessioninfo_1.1.1  classInt_0.4-3     memoise_2.0.0      knitr_1.31
[71] usethis_2.0.1.9000
```

A number of tests are producing warnings. I think this is the same as the issue documented by @mpadge (any ideas on a solution, suppress the warnings in the affected functions?) here: https://github.com/ropensci/osmdata/issues/218#issuecomment-801921318
