### 2021-03-27 #420 @Robinlovelace
Update: finally the test coverage is over 75% (just about!) on the automated coverage check :tada:

![image](https://user-images.githubusercontent.com/1825120/112732435-dbe01f80-8f31-11eb-9440-452dcb5ab73a.png)
