### 2021-03-30 b3e11b7
add rOpenSci badge
terra::rast was not reading the local file (win10). changed to read directly from repo assets.
