+++ README.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/420_status.svg)](https://github.com/ropensci/software-review/issues/420)
 <!-- badges: end -->
 
-The aim of this  Rpackage is to enable fast, accurate and user friendly calculation longitudinal steepness of linear features such as roads and rivers, based on commonly available input datasets such as road geometries and digital elevation model (DEM) datasets.
+This package calculates longitudinal steepness of linear features such as roads and rivers, based on two main inputs: vector linestring geometries and raster digital elevation model (DEM) datasets.
 
 ## Installation
 
@@ -97,8 +97,16 @@ plot(lisbon_road_segments["slope"], add = TRUE, lwd = 5)
 Imagine that we want to go from Santa Catarina to the East of the map to the Castelo de Sao Jorge to the West of the map:
 
 ```{r route}
+library(tmap)
+tmap_mode("view")
+qtm(lisbon_route)
+```
+
+```{r, echo=FALSE, eval=FALSE}
+# Removed because it's not rendering in RMarkdown
 mapview::mapview(lisbon_route)
 ```
+
 
 We can convert the `lisbon_route` object into a 3d linestring object as follows:
 
@@ -143,7 +151,10 @@ For this benchmark we will download the following small (< 100 kB) `.tif` file:
 
 ```{r}
 u = "https://github.com/ITSLeeds/slopes/releases/download/0.0.0/dem_lisbon.tif"
-if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
+f = basename(u)
+if(!file.exists(f)) {
+ download.file(u, f) 
+}
 ```
 
 A benchmark can reveal how many route gradients can be calculated per second:
@@ -151,8 +162,7 @@ A benchmark can reveal how many route gradients can be calculated per second:
 ```{r, results='hide'}
 e = dem_lisbon_raster
 r = lisbon_road_segments
-# et = terra::rast("dem_lisbon.tif")
-et = terra::rast(u)
+et = terra::rast(f)
 res = bench::mark(check = FALSE,
   slope_raster = slope_raster(r, e),
   slope_terra = slope_raster(r, et)
