### 2021-06-09 97909ee
Add slides template
