+++ data-raw/slides.Rmd
@@ -1,10 +1,8 @@
 ---
-title: "Presentation Ninja"
-subtitle: "&#x2694;&#xFE0F; xaringan +<br/>&#x1F60E; xaringanthemer"  
-author: 
-  - "Yihui Xie"
-  - "Garrick Aden-Buie"
-date: '`r Sys.Date()`'
+# title: "slopes: a package for calculating slopes<br>📈🎢🗻🛣️"
+title: "slopes: a package for calculating slopes "
+subtitle: "📈 of roads, rivers and other linear (simple) features 📉"  
+author: "Robin Lovelace & Rosa Félix, UseR 2021<br><br><br><br><br><br><br><br><br><br>"
 output:
   xaringan::moon_reader:
     css: xaringan-themer.css
@@ -15,6 +13,17 @@ output:
       ratio: 16:9
       countIncrementalSlides: true
 ---
+
+```{r, eval=FALSE, echo=FALSE}
+# to run these slides locally:
+xaringan::inf_mr("data-raw/slides.Rmd")
+```
+
+```{r xaringanExtra, echo=FALSE}
+# From https://github.com/gadenbuie/xaringanExtra
+xaringanExtra::use_xaringan_extra(c("tile_view", "animate_css", "tachyons"))
+```
+
 
 ```{r setup, include=FALSE}
 options(htmltools.dir.version = FALSE)
@@ -33,292 +42,61 @@ knitr::opts_chunk$set(
 ```{r xaringan-themer, include=FALSE, warning=FALSE}
 library(xaringanthemer)
 style_duo_accent(
-  primary_color = "#1381B0",
+  title_slide_background_color = "#FFFFFF",
+  title_slide_background_size = "100%",
+  title_slide_background_image = "https://user-images.githubusercontent.com/1825120/121391204-04c75c80-c946-11eb-8d46-ab5d8ada55c2.png",
+  title_slide_background_position = "bottom",
+  title_slide_text_color = "#080808",
+  primary_color = "#080808",
   secondary_color = "#FF961C",
   inverse_header_color = "#FFFFFF"
 )
 ```
 
-## Typography
+background-image: url(https://camo.githubusercontent.com/30a3b814dd72aef5b51db635f2ab6e1b6b6c57b856d239822788967a4932d655/68747470733a2f2f7062732e7477696d672e636f6d2f6d656469612f45724a32647238574d414948774d6e3f666f726d61743d6a7067266e616d653d6c61726765)
+background-position: center
+background-size: 100%
 
-Text can be **bold**, _italic_, ~~strikethrough~~, or `inline code`.
+--
+<br>
+# Structure:
 
-[Link to another slide](#colors).
-
-### Lorem Ipsum
-
-Dolor imperdiet nostra sapien scelerisque praesent curae metus facilisis dignissim tortor. 
-Lacinia neque mollis nascetur neque urna velit bibendum. 
-Himenaeos suspendisse leo varius mus risus sagittis aliquet venenatis duis nec.
-
-- Dolor cubilia nostra nunc sodales
-
-- Consectetur aliquet mauris blandit
-
-- Ipsum dis nec porttitor urna sed
-
----
-name: colors
-
-## Colors
-
-- <span style="color: var(--text-color)">Text Color</span>
-
-- <span style="color: var(--header-color); font-family: var(--header-font-family);">Header Color</span>
-
-- <span style="color: var(--link-color)">Link Color</span>
-
-- <span style="color: var(--text-bold-color); font-weight: bold;">Bold Color</span>
-
-- `inline code color`
-
----
-
-# Big Topic or Inverse Slides `#`
-
-## Slide Headings `##`
-
[…]
