### 2021-06-13 b0775b4
move Benchmark from readme to a proper vignette
