### 2021-06-13 11e1531
Load bench pkg before running benchmark
