+++ vignettes/benchmark.Rmd
@@ -16,6 +16,7 @@ knitr::opts_chunk$set(
 
 ```{r setup}
 library(slopes)
+library(bench)
 ```
 
 # Performance
