### 2021-06-13 41f9b1d
Do not run benchmarking code
