+++ vignettes/benchmark.Rmd
@@ -10,7 +10,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  eval = FALSE
 )
 ```
 
@@ -18,6 +19,8 @@ knitr::opts_chunk$set(
 library(slopes)
 library(bench)
 ```
+
+**Note: benchmarks do not currently evaluate on actions for faster build times.**
 
 # Performance
 
