### 2021-06-13 ITSLeeds/slopes#32 @temospena
Review for ropensci
- [ ] open a parallel branch for revs
- [ ] update readme: include more on the description? the main functions?
- [ ] pass benchmark to a vignete
- [ ] Update argument names
Don't export plot_dz
