+++ man/elevations_get.Rd
@@ -7,9 +7,6 @@
 elevations_get(r, file = NULL, ...)
 }
 \arguments{
-\item{r}{Routes, the gradients of which are to be calculated.
-The object must be of class \code{sf} with \code{LINESTRING} geometries.}
-
 \item{file}{Where to save the resulting data if specified (not implemented)}
 
 \item{...}{Options passed to \code{cc_elevation()}}
