### 2021-06-13 988fa38
Update slope_get with routes not r...
