+++ R/slope_get.R
@@ -7,44 +7,35 @@
 #' @examples
 #' # time-consuming, and see
 #' # https://github.com/ITSLeeds/slopes/runs/648128378#step:18:107
-#' r = cyclestreets_route
-#' # e = elevations_get(r)
+#' routes = cyclestreets_route
+#' # e = elevations_get(routes)
 #' # class(e)
 #' # e
 #' # plot(e)
-#' # plot(sf::st_geometry(r), add = TRUE)
-elevations_get = function(r, file = NULL, ...) {
+#' # plot(sf::st_geometry(routes), add = TRUE)
+elevations_get = function(routes, file = NULL, ...) {
   if(requireNamespace("ceramic")) {
-    mid_ext = sf_mid_ext_lonlat(r)
+    mid_ext = sf_mid_ext_lonlat(routes)
     bw = max(c(mid_ext$width, mid_ext$height)) / 1 # buffer width
     e = ceramic::cc_elevation(loc = mid_ext$midpoint, buffer = bw, ...)
   } else {
     message("Install the package ceramic")
   }
   # issue: cannot convert CRS currently
-  # cr = sf::st_crs(r)
+  # cr = sf::st_crs(routes)
   # cr$wkt
   # raster::crs(e) = raster::crs("+init=epsg:3857")
   # raster::projectRaster(e, "+init=epsg:4326")
-  # raster::projectRaster(e, sf::st_crs(r)[[2]])
+  # raster::projectRaster(e, sf::st_crs(routes)[[2]])
   e
 }
-#
-# trace_projected = sf::st_transform(trace, 3857)
-# plot(e)
-# plot(trace_projected$geometry, add = TRUE)
-#
-# # aim: get max distance from centrepoint
-# bb = sf::st_bbox(sf::st_transform(trace, 4326))
 
-# # max of those 2 and divide by 2
-
-sf_mid_ext_lonlat = function(r) {
+sf_mid_ext_lonlat = function(routes) {
   res = list()
-  if(!sf::st_is_longlat(r)) {
-    r = sf::st_transform(r, 4326)
+  if(!sf::st_is_longlat(routes)) {
+    routes = sf::st_transform(routes, 4326)
   }
-  bb = sf::st_bbox(r)
+  bb = sf::st_bbox(routes)
   res$midpoint = c(mean(c(bb[1], bb[3])), mean(c(bb[2], bb[4])))
   res$width = geodist::geodist(c(x = bb[1], y = bb[2]), c(x = bb[3], y = bb[2]))
   res$height = geodist::geodist(
