+++ R/slopes.R
@@ -20,39 +20,39 @@
 #'
 #' @param x Vector of locations
 #' @param d Vector of distances between points
-#' @param e Elevations in same units as x (assumed to be metres)
+#' @param elevations Elevations in same units as x (assumed to be metres)
 #' @export
 #' @examples
 #' x = c(0, 2, 3, 4, 5, 9)
-#' e = c(1, 2, 2, 4, 3, 1) / 10
-#' slope_vector(x, e)
+#' elevations = c(1, 2, 2, 4, 3, 1) / 10
+#' slope_vector(x, elevations)
 #' m = sf::st_coordinates(lisbon_road_segment)
 #' d = sequential_dist(m, lonlat = FALSE)
-#' e = elevation_extract(m, dem_lisbon_raster)
-#' slope_distance(d, e)
-#' slope_distance_mean(d, e)
-#' slope_distance_weighted(d, e)
-slope_vector = function(x, e) {
+#' elevations = elevation_extract(m, dem_lisbon_raster)
+#' slope_distance(d, elevations)
+#' slope_distance_mean(d, elevations)
+#' slope_distance_weighted(d, elevations)
+slope_vector = function(x, elevations) {
   d = diff(x)
-  e_change = diff(e)
+  e_change = diff(elevations)
   e_change / d
 }
 #' @rdname slope_vector
 #' @export
-slope_distance = function(d, e) {
-  e_change = diff(e)
+slope_distance = function(d, elevations) {
+  e_change = diff(elevations)
   e_change / d
 }
 #' @rdname slope_vector
 #' @export
-slope_distance_mean = function(d, e) {
-  e_change = diff(e)
+slope_distance_mean = function(d, elevations) {
+  e_change = diff(elevations)
   mean(abs(e_change) / d)
 }
 #' @rdname slope_vector
 #' @export
-slope_distance_weighted = function(d, e) {
-  e_change = diff(e)
+slope_distance_weighted = function(d, elevations) {
+  e_change = diff(elevations)
   stats::weighted.mean(abs(e_change) / d, d)
 }
 #' Calculate the gradient of line segments from a 3D matrix of coordinates
@@ -74,16 +74,16 @@ slope_distance_weighted = function(d, e) {
 #' points(x[-length(x)], gxy, col = "blue")
 #' title("Distance (in x coordinates) elevation profile",
 #'   sub = "Points show calculated gradients of subsequent lines")
-slope_matrix = function(m, e = m[, 3], lonlat = TRUE) {
+slope_matrix = function(m, elevations = m[, 3], lonlat = TRUE) {
   d = sequential_dist(m, lonlat = lonlat)
-  e_change = diff(e)
+  e_change = diff(elevations)
   g = e_change / d
   g
 }
 #' @rdname slope_matrix
 #' @export
-slope_matrix_weighted = function(m, e = m[, 3], lonlat = TRUE) {
-  g1 = slope_matrix(m, e = e, lonlat = lonlat)
+slope_matrix_weighted = function(m, elevations = m[, 3], lonlat = TRUE) {
+  g1 = slope_matrix(m, elevations = elevations, lonlat = lonlat)
   d = sequential_dist(m = m, lonlat = lonlat)
   stats::weighted.mean(abs(g1), d, na.rm = TRUE)
 }
@@ -137,7 +137,7 @@ slope_matrices = function(m_xyz_split, fun = slope_matrix_weighted, ...) {
 #' @inheritParams elevation_extract
 #' @param routes Routes, the gradients of which are to be calculated.
 #'   The object must be of class `sf` with `LINESTRING` geometries.
-#' @param e Raster overlapping with `routes` and values representing elevations
[…]
