+++ R/slopes.R
@@ -90,7 +90,11 @@ slope_matrix_weighted = function(m, elevations = m[, 3], lonlat = TRUE) {
 
 #' Calculate the sequential distances between sequential coordinate pairs
 #'
-#' @param lonlat Are the coordinates in lon/lat order? `TRUE` by default
+#' @param lonlat Are the coordinates in lon/lat (geographic) coordinates?
+#'  Set this to `FALSE` if you have projected data, e.g. with coordinates
+#'  representing distance in meters, not degrees. Lonlat coodinates are assumed
+#'  (`lonlat = TRUE`) is the default. In `slope_raster()` and `plot_slope()`
+#'  it is taken from the CRS of the routes (`sf::st_is_longlat(routes)`).
 #' @inheritParams slope_matrix
 #' @export
 #' @examples
@@ -177,27 +181,19 @@ slope_raster = function(
 #' @export
 #' @examples
 #' r_xyz = lisbon_road_segment_3d
-#' slope_xyz(r_xyz)
-slope_xyz = function(
-  r_xyz,
-  fun = slope_matrix_weighted,
-  lonlat = sf::st_is_longlat(r_xyz)
-  ) {
-  # browser()
+#' slope_xyz(r_xyz, lonlat = FALSE)
+slope_xyz = function(r_xyz, fun = slope_matrix_weighted, lonlat = TRUE) {
   if(inherits(r_xyz, "sf") | inherits(r_xyz, "sfc")) {
-    if(is.null(lonlat)) {
-      lonlat = sf::st_is_longlat(r_xyz)
-    }
+    lonlat = sf::st_is_longlat(r_xyz)
     r_xyz = as.data.frame(sf::st_coordinates(r_xyz))
-  }
-  if(is.null(lonlat)) {
-    lonlat = FALSE
   }
   if("L1" %in% colnames(r_xyz)) {
     m_xyz_split = split(x = r_xyz, f = r_xyz[, "L1"])
     res = slope_matrices(m_xyz_split, lonlat = lonlat, fun = fun)
   } else {
-    # todo: add content here
+    # todo: add content here if data frame was generated by sfheaders
+    # or another package that does not call id colums 'L1' by default
+    # (low priority)
   }
   res
 }
