+++ man/sequential_dist.Rd
@@ -9,7 +9,11 @@ sequential_dist(m, lonlat = TRUE)
 \arguments{
 \item{m}{Matrix containing coordinates and elevations}
 
-\item{lonlat}{Are the coordinates in lon/lat order? \code{TRUE} by default}
+\item{lonlat}{Are the coordinates in lon/lat (geographic) coordinates?
+Set this to \code{FALSE} if you have projected data, e.g. with coordinates
+representing distance in meters, not degrees. Lonlat coodinates are assumed
+(\code{lonlat = TRUE}) is the default. In \code{slope_raster()} and \code{plot_slope()}
+it is taken from the CRS of the routes (\code{sf::st_is_longlat(routes)}).}
 }
 \description{
 Calculate the sequential distances between sequential coordinate pairs
