### 2021-06-13 a9335bb
create vignette for example with Isle of Wight
