+++ vignettes/roadnetworkcycling.Rmd
@@ -1,0 +1,191 @@
+---
+title: "Example: gradients of a road network for a given city"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Example: gradients of a road network for a given city}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.width=7, fig.height=5, fig.align = "center" 
+)
+```
+
+An example of the demand for data provided by the `slopes` package is a map showing gradients across São Paulo (Brazil, see image below), with a simplistic classification for cycling difficulty.
+
+![using slopes() to create a road network gradient for cycling for São Paulo (Brazil)](https://pbs.twimg.com/media/ErJ2dr8WMAIHwMn?format=jpg&name=small)
+
+This vignette will guide through the production of an interactive slope map for a road network, using `slopes`, `osmextract`, `sf`, `stplanr` and `tmap`.  
+
+For the convenience of sample, we will use (Isle of Wight)[https://en.wikipedia.org/wiki/Isle_of_Wight] case, with 384 km2. See [Other examples] below.
+
+This will follow three steps:
+
+1.  Download of road network from [OpenStreetMap](https://www.openstreetmap.org/)
+2.  Prepare the network
+3.  Compute slopes and export the map in html
+
+
+## Extract the OSM network from geofabrik
+
+For this step you may use `osmextract` [package](https://itsleeds.github.io/osmextract/articles/osmextract.html) which downloads the most recent information available at OSM (https://download.geofabrik.de/index.html) and converts to _GeoPackage_ (.gpkg), the equivalent to _shapefile_.
+
+```{r setup1, warning=FALSE, message = FALSE}
+library(dplyr)
+library(sf)
+# remotes::install_github("ITSLeeds/osmextract")
+library(osmextract)
+```
+
+```{r get_iow, warning=FALSE, message = FALSE}
+# get the network
+iow_osm = oe_get("Isle of Wight", provider = "geofabrik", stringsAsFactors = FALSE, 
+                 quiet = FALSE, force_download = TRUE, force_vectortranslate = TRUE) # 7 MB
+
+# filter the major roads
+iow_network = iow_osm %>% 
+  dplyr::filter(highway %in% c('primary', "primary_link", 'secondary',"secondary_link", 
+                               'tertiary', "tertiary_link", "trunk", "trunk_link", 
+                               "residential", "cycleway", "living_street", "unclassified", 
+                               "motorway", "motorway_link", "pedestrian", "steps", "track")) #remove: "service"
+```
+
+## Clean the road network
+
+These are optional steps that give better results, although they may slow down the process since they increase the number of segments present in the network.
+
+### Filter the unconnected segments
+
+The [`rnet_group()`](https://docs.ropensci.org/stplanr/reference/rnet_group.html) function from `stplanar` package assesses the connectivity of each segment assigns a group number (similar to a clustering process). Then we may filter the main group, the one with more connected segments.
+
+```{r setup2, warning=FALSE, message = FALSE}
+# remotes::install_github("ropensci/stplanr")
+library(stplanr)
+```
+
+```{r filter}
+# filter unconnected roads
+iow_network$group = rnet_group(iow_network)
+iow_network_clean = iow_network %>% filter(group == 1) #the network with more connected segments
+```
+
+### Break the segments on vertices
+
[…]
