### 2021-06-13 #420 @temospena
Thank you very much for your excellent comments!
We have started the [revision process](https://github.com/ITSLeeds/slopes/pull/31) and will let you know as soon as it is completed.
