### 2021-06-13 27bc145
Dont run unexported function (plot_dz)
