+++ R/plot_slope.R
@@ -54,7 +54,8 @@ plot_slope = function(routes,
 #' d = cumsum(sequential_dist(m, lonlat = FALSE))
 #' d = c(0, d)
 #' z = m[, 3]
-#' plot_dz(d, z)
+#' # not exported
+#' # plot_dz(d, z)
 plot_dz = function(d,
                    z,
                    fill = TRUE,
