+++ man/plot_dz.Rd
@@ -62,5 +62,6 @@ m = sf::st_coordinates(routes)
 d = cumsum(sequential_dist(m, lonlat = FALSE))
 d = c(0, d)
 z = m[, 3]
-plot_dz(d, z)
+# not exported
+# plot_dz(d, z)
 }
