### 2021-06-13 730180c
Dont evaluage roadnetwork cycling vignette for actions
