+++ vignettes/roadnetworkcycling.Rmd
@@ -11,7 +11,8 @@ vignette: >
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.width=7, fig.height=5, fig.align = "center" 
+  fig.width=7, fig.height=5, fig.align = "center",
+  eval = FALSE
 )
 ```
 
