### 2021-06-13 caa12a3
Try hardcoding values (is there a better way?
