### 2021-06-13 ITSLeeds/slopes#32 @Robinlovelace
Image for new vignette

![image](https://user-images.githubusercontent.com/1825120/121820334-2435f080-cc8a-11eb-962c-79dcba97e459.png)
