### 2021-06-13 ITSLeeds/slopes#32 @Robinlovelace
This is a great map but it adds ~13 MB to the package size as it's embedded in the html of the vignette. Plan: add link to interactive version.
