### 2021-06-13 18e4950
Add stplanr and dplyr as Suggests
