### 2021-06-13 ITSLeeds/slopes#32 @temospena
Perfect, that is a much better idea!

Should I put it on branch rev2-fix?
