### 2021-06-13 ITSLeeds/slopes#32 @temospena
Just saw that you already did that on
https://github.com/ITSLeeds/slopes/commit/1942e94f9f7c75948004ef9cf225e2e8911792ea
:)
