### 2021-06-14 fd19706
add rgdal to description and rebuild
