### 2021-06-21 ITSLeeds/slopes#32 @Robinlovelace
Draft response to review 2:

> ## Package Review
>
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
>     * **Briefly describe any working relationship you have (had) with the package authors.**
>
>     * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
>
> #### Documentation
>
> The package includes all the following forms of documentation:
>
>     * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
>
>     * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
>
>     * [ ]  **Vignette(s)** demonstrating major functionality that runs successfully locally
>
>     * [x]  **Function Documentation:** for all exported functions
>
>     * [ ]  **Examples** (that run successfully locally) for all exported functions
>
>     * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
>
> #### Functionality
>
>     * [x]  **Installation:** Installation succeeds as documented.
>
>     * [x]  **Functionality:** Any functional claims of the software been confirmed.
>
>     * [x]  **Performance:** Any performance claims of the software been confirmed.
>
>     * [x]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
>     * [ ]  **Packaging guidelines:**: The package conforms to the rOpenSci packaging guidelines
>
>
> Estimated hours spent reviewing: 8-10
>
>     * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
>
>
> ## Review Comments
>
> This is a great package that fills an obvious need. It has all of the necessary functionality required to do the job it sets out to do - calculate slopes of routes. The functions for getting and using the appropriate DEM for an area removes a huge barrier for users, so kudos! I had fun test-driving the package and learned a lot. I hope my review doesn't come across as negative - it really is an excellent package and with some rough edges filed down I think it could be even better! I did get a bit in the weeds at times - so while there are a lot of comments most of them are quite small!
> ## Review Checklist
> ### Documentation
> #### Statement of Need
>
> The introductory sentence in the README gives a very brief description of what the package does, and the target audience is fairly easily inferred (i.e., people who need to calculate slopes of linear features). I don't think this needs much more as the application is fairly clear, however you could include an overview of some of the other things the package does to help a user get to the point of calculating slopes (e.g., getting elevation data from online sources, extracting elevation points), as well as the visualization tools (plotting functions). I like that it compares the package's functionality to other common tools (ESRI's `Slope3d()`).
> #### Installation instructions
>
>     * Installation instructions work and are straightforward, however it would be helpful to add some information about MapBox, and that an API key is needed for some functions. Also note that for some functions, there are packages that will need to be installed separately as they are not installed automatically (i.e., are in `Suggests` - `ceramic`, `ggdist`, etc.)
>
>
> #### Vignette(s)
>
>     * The main vignette (`slopes.Rmd`) doesn't actually show any examples of what the stated main purpose of the package is - to provide slopes of linear features from a DEM. The README has the most comprehensive walkthrough of the main advertised functionality of the package. It is available on GitHub and the pkgdown page - but not available in R as a vignette. I suggest taking the README content and creating a vignette (likely the primary one), or perhaps even better - add it to the current `slopes.Rmd` so you combine the theory with a demonstration of the implementation.
>
>     * `verification.Rmd`:
>
>       * a `library(raster)` call is needed to plot the raster object (`plot(e)` on [line 91](https://github.com/ITSLeeds/slopes/blob/3a25f7ce11ced3be313c0eadbda395551887fa2e/vignettes/verification.Rmd#L91))
>       * `slope_raster(r, e = e)` on [line 123](https://github.com/ITSLeeds/slopes/blob/3a25f7ce11ced3be313c0eadbda395551887fa2e/vignettes/verification.Rmd#L123) fails because `points2line_trajectory()` does not assign CRS to the output - as a result, `lonlat` is not determined internally (more on this below). I made it work here by tweaking `points2line_trajectory()`:
>
>       ```
>       points2line_trajectory = function(p) {
>         c = st_coordinates(p)
>         i = seq(nrow(p) - 2)
>         l = purrr::map(i, ~ sf::st_linestring(c[.x:(.x + 1), ]))
>         lfc = sf::st_sfc(l)
>         a = seq(length(lfc)) + 1 # sequence to subset
>         p_data = cbind(sf::st_set_geometry(p[a, ], NULL))
>         sf::st_sf(p_data, geometry = lfc, crs = st_crs(p)) # set the crs here
>       }
>       ```
>
>
> #### Function Documentation
>
> **General Comments:**
>
>     * Package-level documentation would be convenient, so a user can type `?slopes` and get an overview of what the package does, authors, urls etc. - and link to the package index). You can set this up with `usethis::use_package_doc()`.
>
>     * All function documentation should have a `@value` (return) section so the user knows what they can expect from the function.
>
>     * The included datasets are great - nice tidy real-world examples to play with.
>
>     * The datasets could be documented a bit more clearly - how many road segments, lengths/distances etc. It's nice to have the context when you are playing with the examples. It is not obvious to the user the differences between the various datasets with Lisbon road segments.
>
>     * In a lot of cases the names of function arguments could be clearer - many are single letters that don't have an obvious meaning.
>
>     * Data documentation should provide clearer sources for the data - for example, `dem_lisbon_raster` has only a link to a `terra` GitHub issue. The various Lisbon road segment datasets link to the ESRI's 3D Analyst extension webpage - which is not the source of the data, rather the tool that created it.
>
>
> **Specific documentation issues:**
>
>     * `elevations_get()` could use more explanation of how these elevations are obtained and how to customize it. What web service(s) is it using? I see that it uses MapBox via `ceramic::cc_elevation()` however that might be a bit opaque to a user. I think at least a mention that they need a MapBox API key (ideally with a link/instructions on how to get one), and a link to the `ceramic::cc_elevation()` documentation would be helpful, as well as some examples showing customization by passing options to `cc_elevation()`.
>
>     * `elevation_extract()` documentation would benefit from more explanation of its arguments: what other options can be used for `method`? What are the benefits of using `terra` (and what is the alternative? - I know it's `raster` but the user may not).
>
>     * `plot_dz`:
>
>       * All argument documentation should be expanded, at a minimum stating the defaults, the accepted options (and/or type/format of inputs).
>       * I tend to think that the [setting of the default colour](https://github.com/ITSLeeds/slopes/blob/9365709a95312b4feff3689a9e219c06af314167/R/plot_slope.R#L63-L66) palette (`p`) could be done internally in the function, and set the default argument value to `NULL` - this would be a simpler interface for the user. In addition, the value of `requireNamespace()` does not need to be assigned to anything in this `ifelse()` statement.
>       * It is not clear (to me at least) what the argument `x` does - I can infer that it is legend position, but neither the name of the argument nor the description tell you this. I suggest renaming it and explaining what it does.
>       * It's not clear what the values passed to `s` do, nor what significance the default values `3:18` hold (especially as they relate to `brks`).
>
>     * `plot_slope()` - Similar comments on argument names, defaults and descriptions as `plot_dz`. It seems to me that this is the function that users would call much more often than manually constructing the components and then using `plot_dz()` - if that's the case, would you consider not exporting `plot_dz`, or directing users to `plot_slope()` in most cases?
>
>     * `sequential_dist()` - Some more clarity on what `m` should look like would be helpful. What are the dimensions of the matrix, what are the order of the columns? The example doesn't give much clarity on this.
>
>     * `slope_3d()` - it is not clear from the documentation that if you omit the raster object (`e = NULL`) then the function will download the necessary raster from MapBox - this is amazing functionality and should be advertised!
>
>     * The documentation links in `?slope_vector` are not necessary since they link to the same help page.
>
>     * I anticipated that there would be a `slope_matrix_mean()` that calculates the unweighted mean slope from a matrix, in addition to `slope_matrix()` and `slope_matrix_weighted()` - analogous to `slope_distance_mean()`.
>
>     * `slope_xyz` - Alternative functions that can be passed to `fun` should be documented.
>
>
> #### Examples
>
> **General Comments:**
>
>     * All functions have examples that show the basic usage.
>
>     * I think it would be good to standardize how functions are called in examples (sometimes they are called with `package::function()` and sometimes just `function()`). I think for `slopes` functions you probably don't need to use the `::` notation, as it assumed the package is loaded, and for functions from other packages you should standardize on either using (for example) `library(sf)` or `sf::function()` (I would lean towards the former to avoid the method dispatch issues I flagged in several examples below).
>
>     * In general, I think the examples could contain a bit more explanation of what is happening and/or use more meaningful values in the examples (e.g., real lon/lat pairs and distances rather than simple integers) - this would facilitate understanding the intention and outputs.
>
>     * Rather than commenting out examples that should not be automtically run, it would be better to wrap them in `\dontrun{}` or `\donttest{}`.
>
>
> **Specific Issues:**
>
>     * In the example for `cyclestreets_route`, I believe this contains the (commented out) code for generating the data, not an example of its use. I think this could be confusing for the user, and should probably belong in a script in the `data-raw` folder - the code also uses the `stplanr` package, which is not listed as a dependency in `slopes`. In order to run the commented-out `plot()` call, the `sf` package must be loaded to use the correct `plot()` method.
>
>     * In the `dem_lisbon_raster()` example the `sf` package must be loaded to use the correct `plot()` method.
>
>     * In the example for `slope_raster()`, the `sf` package must be loaded for the correct `[` method to be used (otherwise `r = lisbon_road_segments[1:3, ]` results in an invalid `sf` object).
>
>
> #### Community guidelines
>
>     * DESCRIPTION file has URLs for GitHub and pkgdown pages, and a BugReports field +1
>
>     * There is currently no CONTRIBUTING file or instructions on how to contribute in the README.
>
>     * Currently there is no Code of Conduct, though it appears that will be [added upon acceptance to rOpenSci](https://devguide.ropensci.org/collaboration.html?q=conduct#coc-file).
>
>
> ### Functionality
>
> The package ticks almost all of the boxes in the _functionality_ checklist - Installation was quick and easy (I'm on a Mac), functional claims are satisfied, and performance was as expected. It has automated tests (see a few comments below), but there are a few parts of the rOpenSci packaging guidelines which are not met.
> #### Automated Tests
>
>     * Good test coverage (75%) though more tests that check edge cases and anticpated common user mistakes would be beneficial.
>
>     * Unconditional use of Suggested packages `geodist` and `colorspace` in tests could cause problems if they are not automatically installed.
>
>
> #### rOpenSci Packaging Guidelines
>
>     * The package name is perfect.
>
>     * There is no CodeMeta file.
>
>     * It runs great on my Mac, and I see GitHub Actions is set up to check on all major platforms. R CMD check passed successfully (and quickly) locally on my machine.
>
>     * Please see other comments on API, function and argument naming, documentation, examples, testing, etc.
>
>     * Dependencies: There are quite a few packages in `Suggests:`. I definitely agree with reducing the number hard dependencies, I do wonder if some of these are important enough to be moved into `Imports` - I am specifically thinking about `geodist` and `pbapply`.
>
>     * The pkgdown site is great.
>
>
> ## General Code Comments
>
> The code generally follows good practices - it's legible and well structured, with little duplication (other than exceptions mentioned below). I think most functions would benefit from more thorough argument checking so that a user gets an early and informative error message if they try to use inappropriate argument values.
>
> There are commented-out blocks of code sprinkled throughout, where the intent of them is not clear. Perhaps clarify with comments/TODOs and/or remove where appropriate.
> ### Specific Code Comments
>
>     * Very nice use of diverging colour palettes in plots for positive/negative slopes.
>
>     * `elevations_get()` - It looks like you are planning to implement caching functionality, which would be great. I think at the very least it would be worth enabling saving to disk via the `file` argument (currently marked as not implemented).
>
>     * `plot_slope()` and `slope_raster()`: `lonlat` cannot be determined with `sf::st_is_longlat(r)` if `r` has no CRS as it returns `NA`, resulting in an error deeper in the stack. This should be caught early and an informative error should be thrown.
>
>     * `pbapply` is used in `slope_matrices()`, a major workhorse function - I think it should be in `Imports` (and then no need to use `requireNamespace`).
>
>     * `elevation_extract()` works (perhaps unintentionally) directly with `sf` objects, however it returns different results than using it on the component coordinates, as illustrated in the examples. When done with an `sf` object it appears to return a list containing a vector of elevations for each line segment - which I think could be really useful! If you want this functionality I would recommend creating an S3 generic + methods and documenting the differences.
>       ```
>       m = sf::st_coordinates(lisbon_road_segments[1, ])
>       e = dem_lisbon_raster
>       elevation_extract(m, e)
>       #>  [1] 3.267991 3.266818 3.265715 3.264418 3.263222 3.262339 3.261004 3.262972
>       #>  [9] 3.264940 3.266908 3.268876 3.308304 3.461349 3.467100 3.439341 3.421206
>       #> [17] 3.403916 3.387750 3.371584 3.356889 3.342193 3.327498 3.312644 3.297117
>       #> [25] 3.282411 3.267716 3.253020 3.218922 3.210919 3.202916 3.194914 3.176672
>       #> [33] 3.157027 3.191458 3.316882 3.410792 3.469193 3.487912 3.563260 3.640330
>       #> [41] 3.717938
>       elevation_extract(lisbon_road_segments[1, ], e)
>       #> Warning in showSRID(uprojargs, format = "PROJ", multiline = "NO", prefer_proj
>       #> = prefer_proj): Discarded datum Unknown based on GRS80 ellipsoid in Proj4
>       #> definition
>       #> Warning in showSRID(SRS_string, format = "PROJ", multiline = "NO", prefer_proj
>       #> = prefer_proj): Discarded datum European Terrestrial Reference System 1989 in
>       #> Proj4 definition
>       #> [[1]]
>       #>  [1] 3.370 3.382 3.394 3.406 3.419 3.431 3.104 3.116 3.128 3.140 3.152 3.233
>       #> [13] 3.507 3.400 3.389 3.374 3.363 3.352 3.341 3.330 3.319 3.308 3.296 3.285
>       #> [25] 3.274 3.263 3.252 3.241 3.230 3.219 3.208 3.315 3.451 3.560 3.433 3.543
>       #> [37] 3.652 3.762
>       ```
>
>     * `slope_z_*()`: In `slope_z_mean()`, `mean()` is called with `na.rm = TRUE`, however `min()` and `max()` are not in their respective `slope_x_` functions - I recommend standardizing that, and/or exposing that argument to the `slope_z_*()` functions.
>
>     * Many functions that say they must be used with `sf` objects could also be used with `sfc` and even `sfg` objects. It might be a good idea to support this, possibly by making a generic with methods for each? Or in many cases it may be fine to just rely on the S3 methods from the `sf` functions (e.g., `sf::st_coordinates()` has methods for `sf`, `sfc`, `sfg` and should "just work"). If this isn't a desirable direction, there should be more comprehensive checking of inputs to ensure good error messages if `r` is not `sf`.
>
>     * There is lots of code duplication in `z.R` - the internals of most of these could be abstracted out to one or two functions. In addition, using the `sf::st_z_range()` function might be useful here.
>
>     * If the `Suggested` package `colorspace` is not installed, `plot_slope()` and `plot_dz()` will fail, due to the package being [called unconditionally in `make_pal()`](https://github.com/ITSLeeds/slopes/blob/9365709a95312b4feff3689a9e219c06af314167/R/plot_slope.R#L130).
>
>     * `plot_slope()` and `plot_dz()` - most of the aesthetic arguments in `plot_slope()` are not passed on to `plot_dz()` thus have no effect.
>
>     * `elevations_get()` - You transform the input `sf` object to lonlat to pass through to MapBox, but currently return the return the raster in its native CRS (Web Mercator I believe). This may not match the user's input, so they will likely have to transform one or the other for them to work together. This is probably fine, but the user should probably be given a message or a warning to advise them. Likewise, in `slope_3d()` it would be good to check that the CRSs of the two datasets match and throw an error otherwise.
>
>     * In `slope_3d()` much of the code is duplicated and could be pulled out into small helper functions (in both conditions after checking `is.null(e)` on [line 261](https://github.com/ITSLeeds/slopes/blob/c018c5d65208728cea66a1a13451a4c568deef5f/R/slopes.R#L261) and in both conditions after [checking nrow()](https://github.com/ITSLeeds/slopes/blob/c018c5d65208728cea66a1a13451a4c568deef5f/R/slopes.R#L279)). Additionally the comment on [line 2810](https://github.com/ITSLeeds/slopes/blob/c018c5d65208728cea66a1a13451a4c568deef5f/R/slopes.R#L280) seems to be outdated since it does work with n > 1.
>
>
> **Function Argument Naming**:
>
> Function argument names `e`, `r`, and `lonlat` are used across multiple functions for different types, sometimes with different defaults and descriptions. I suggest standardizing these and using different names where appropriate. In general, I think preferring longer, informative argument names over single letters would be helpful. I created a table of the different ways these arguments are used and/or documented:
> Arg Name 	Function 	Arg Description 	Default value 	Notes
> `r` 	`elevations_get` 	Routes, the gradients of which are to be calculated. The object must be of class `sf` with `LINESTRING` geometries. 	None
[…]
