+++ DESCRIPTION
@@ -30,12 +30,12 @@ RoxygenNote: 7.1.1
 Imports: 
     sf,
     raster,
-    methods
+    methods,
+    pbapply
 Depends: 
     R (>= 2.10)
 Suggests: 
     geodist,
-    pbapply,
     terra,
     colorspace,
     knitr,
