+++ pkgdown.yml
@@ -2,7 +2,9 @@ pandoc: 2.7.3
 pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
+  benchmark: benchmark.html
+  roadnetworkcycling: roadnetworkcycling.html
   slopes: slopes.html
   verification: verification.html
-last_built: 2021-06-13T10:16Z
+last_built: 2021-06-21T08:00Z
 
