+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   roadnetworkcycling: roadnetworkcycling.html
   slopes: slopes.html
   verification: verification.html
-last_built: 2021-06-23T21:03Z
+last_built: 2021-06-24T14:08Z
 
