### 2021-06-28 #420 @annakrystalli
Hello @temospena ! Just wanted to check in with you about how the revisions are going.

Feel free to ask for any clarification if you are unsure of anything at any point.
