### 2021-07-04 b12f981
Update slides with ground truth figure
