### 2021-07-04 df0b892
Update slides - no research bullet point
