### 2021-07-04 4826f1d
Add final worked example
