### 2021-08-02 #420 @Robinlovelace
Thanks for the reminder, holidays and other deadlines have slowed down progress on this but progress there is. See here for a merged PR that tackles many of the issues reported by reviewer 2: https://github.com/ITSLeeds/slopes/pull/31

More soon-ish... there is lots of value in the comments so committed to acting on them :sparkles:
