+++ R/plot_slope.R
@@ -14,11 +14,7 @@ plot_slope = function(routes,
                       lonlat = sf::st_is_longlat(routes),
                       fill = TRUE,
                       horiz = FALSE,
-                      p = ifelse(
-                        test = requireNamespace("colorspace", quietly = TRUE),
-                        colorspace::diverging_hcl,
-                        grDevices::terrain.colors
-                      ),
+                      p = colorspace::diverging_hcl,
                       ...,
                       x = "top",
                       col = "black",
@@ -60,11 +56,7 @@ plot_dz = function(d,
                    z,
                    fill = TRUE,
                    horiz = FALSE,
-                   p = ifelse(
-                     test = requireNamespace("colorspace", quietly = TRUE),
-                     colorspace::diverging_hcl,
-                     grDevices::terrain.colors
-                   ),
+                   p =  colorspace::diverging_hcl,
                    ...,
                    x = "top",
                    col = "black",
