+++ man/plot_dz.Rd
@@ -9,8 +9,7 @@ plot_dz(
   z,
   fill = TRUE,
   horiz = FALSE,
-  p = ifelse(test = requireNamespace("colorspace", quietly = TRUE),
-    colorspace::diverging_hcl, grDevices::terrain.colors),
+  p = colorspace::diverging_hcl,
   ...,
   x = "top",
   col = "black",
