+++ man/plot_slope.Rd
@@ -9,8 +9,7 @@ plot_slope(
   lonlat = sf::st_is_longlat(routes),
   fill = TRUE,
   horiz = FALSE,
-  p = ifelse(test = requireNamespace("colorspace", quietly = TRUE),
-    colorspace::diverging_hcl, grDevices::terrain.colors),
+  p = colorspace::diverging_hcl,
   ...,
   x = "top",
   col = "black",
