### 2021-08-17 7ff342a
rename elevations_get() to elevation_get() - singular for consistency. #32
