+++ tests/testthat/test-get.R
@@ -2,7 +2,7 @@ test_that("functions to get elevations from sf objects work", {
   if(nchar(Sys.getenv("MAPBOX_API_KEY")) < 8)
     skip(message = "Skipping test, MAPBOX token in .Renviron needed")
   r = cyclestreets_route
-  e = elevations_get(r)
+  e = elevation_get(r)
   expect_equal(class(e)[1], "RasterLayer")
 })
 
