+++ man/elevation_get.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/slope_get.R
-\name{elevations_get}
-\alias{elevations_get}
+\name{elevation_get}
+\alias{elevation_get}
 \title{Get elevation data from hosted maptile services}
 \usage{
-elevations_get(routes, file = NULL, ...)
+elevation_get(routes, file = NULL, ...)
 }
 \arguments{
 \item{routes}{Routes, the gradients of which are to be calculated.
@@ -21,7 +21,7 @@ Get elevation data from hosted maptile services
 # time-consuming, and see
 # https://github.com/ITSLeeds/slopes/runs/648128378#step:18:107
 routes = cyclestreets_route
-# e = elevations_get(routes)
+# e = elevation_get(routes)
 # class(e)
 # e
 # plot(e)
