+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(elevation_extract)
-export(elevations_get)
+export(elevation_get)
 export(plot_slope)
 export(sequential_dist)
 export(slope_3d)
