### 2021-08-17 877b187
rename `slope_3d()` to `elevation_add()` for a more intuitive funcion name and consistency. #32
