+++ README.Rmd
@@ -92,7 +92,7 @@ summary(lisbon_road_segments$slope)
 This created a new column, `slope` that represents the average, distance weighted slope associated with each road segment.
 The units represent the percentage incline, that is the change in elevation divided by distance.
 The summary of the result tells us that the average gradient of slopes in the example data is just over 5%.
-This result is equivalent to that returned by ESRI's `Slope_3d()` in the [3D Analyst extension](https://desktop.arcgis.com/en/arcmap/10.3/tools/3d-analyst-toolbox/slope.htm), with a correlation between the ArcMap implementation and our implementation of more than 0.95 on our test dataset (we find higher correlations on larger datasets):
+This result is equivalent to that returned by ESRI's `elevation_add()` in the [3D Analyst extension](https://desktop.arcgis.com/en/arcmap/10.3/tools/3d-analyst-toolbox/slope.htm), with a correlation between the ArcMap implementation and our implementation of more than 0.95 on our test dataset (we find higher correlations on larger datasets):
 
 ```{r}
 cor(
@@ -137,13 +137,13 @@ nodes = stplanr::find_network_nodes(sln, coords[, 1], coords[, 2])
 lisbon_route = stplanr::sum_network_routes(sln = sln, start = nodes[1], end = nodes[2])
 mapview::mapview(lisbon_route) +
   mapview::mapview(lisbon_road_segments["slope"])
-lisbon_route_3d = slope_3d(lisbon_route, dem_lisbon_raster)
+lisbon_route_3d = elevation_add(lisbon_route, dem_lisbon_raster)
 usethis::use_data(lisbon_route, overwrite = TRUE)
 usethis::use_data(lisbon_route_3d, overwrite = TRUE)
 ```
 
 ```{r}
-lisbon_route_3d = slope_3d(lisbon_route, dem_lisbon_raster)
+lisbon_route_3d = elevation_add(lisbon_route, dem_lisbon_raster)
 ```
 
 We can now visualise the elevation profile of the route as follows:
@@ -155,7 +155,7 @@ plot_slope(lisbon_route_3d)
 If you do not have a raster dataset representing elevations, you can automatically download them as follows.
 
 ```{r, message=FALSE, warning=FALSE}
-lisbon_route_3d_auto = slope_3d(lisbon_route)
+lisbon_route_3d_auto = elevation_add(lisbon_route)
 plot_slope(lisbon_route_3d_auto)
 ```
 
