+++ man/elevation_add.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/slopes.R
-\name{slope_3d}
-\alias{slope_3d}
+\name{elevation_add}
+\alias{elevation_add}
 \title{Take a linestring and add a third (z) dimension to its coordinates}
 \usage{
-slope_3d(
+elevation_add(
   routes,
   dem = NULL,
   method = "bilinear",
@@ -31,11 +31,11 @@ Take a linestring and add a third (z) dimension to its coordinates
 \examples{
 routes = lisbon_road_segments[204, ]
 dem = dem_lisbon_raster
-(r3d = slope_3d(routes, dem))
+(r3d = elevation_add(routes, dem))
 sf::st_z_range(routes)
 sf::st_z_range(r3d)
 plot(sf::st_coordinates(r3d)[, 3])
 plot_slope(r3d)
-# r3d_get = slope_3d(cyclestreets_route)
+# r3d_get = elevation_add(cyclestreets_route)
 # plot_slope(r3d_get)
 }
