+++ NAMESPACE
@@ -1,10 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
+export(elevation_add)
 export(elevation_extract)
 export(elevation_get)
 export(plot_slope)
 export(sequential_dist)
-export(slope_3d)
 export(slope_distance)
 export(slope_distance_mean)
 export(slope_distance_weighted)
