### 2021-08-17 44ee193
rename all slope_z_..() funcions to z_...() for more intuitive names. leave "slope_.." for funcions that deal with slope values - not elevation. #32
