+++ tests/testthat/test-z.R
@@ -1,13 +1,13 @@
 test_that("Functions on Z values work", {
   x = slopes::lisbon_route_3d
-  expect_equal(class(slope_z_value(x)), "numeric")
+  expect_equal(class(z_value(x)), "numeric")
   x = slopes::lisbon_route
-  expect_error(slope_z_value(x))
-  expect_error(slope_z_start(x))
-  expect_error(slope_z_end(x))
-  expect_error(slope_z_mean(x))
-  expect_error(slope_z_max(x))
-  expect_error(slope_z_min(x))
+  expect_error(z_value(x))
+  expect_error(z_start(x))
+  expect_error(z_end(x))
+  expect_error(z_mean(x))
+  expect_error(z_max(x))
+  expect_error(z_min(x))
   # x_slope =
   # expect_is()
 })
