### 2021-08-17 f9de5cc
New function arg route_xyz
