### 2021-08-17 2e52d67
Update docs with devtools::document()
