+++ NAMESPACE
@@ -13,10 +13,10 @@ export(slope_matrix_weighted)
 export(slope_raster)
 export(slope_vector)
 export(slope_xyz)
-export(slope_z_end)
-export(slope_z_max)
-export(slope_z_mean)
-export(slope_z_min)
-export(slope_z_start)
-export(slope_z_value)
+export(z_end)
+export(z_max)
+export(z_mean)
+export(z_min)
+export(z_start)
+export(z_value)
 importFrom(methods,is)
