### 2021-08-17 7d9bd12
Load sf library, fix CI (hopefully)
