+++ man/elevation_add.Rd
@@ -29,6 +29,7 @@ if \code{dem} is of class \code{SpatRast}}
 Take a linestring and add a third (z) dimension to its coordinates
 }
 \examples{
+library(sf)
 routes = lisbon_road_segments[204, ]
 dem = dem_lisbon_raster
 (r3d = elevation_add(routes, dem))
