### 2021-08-17 ef15955
add `na.rm=` to z_min() and z_max() fopr consistency. #32
