### 2021-08-17 a3af4cf
Add code of conduct and contributing. #32
