+++ README.Rmd
@@ -160,4 +160,9 @@ plot_slope(lisbon_route_3d_auto)
 ```
 
 
+## Code of Conduct
+
+Please note that the slopes project is released with a [Contributor Code of Conduct](https://contributor-covenant.org/version/2/0/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
+
+
 
