### 2021-08-17 7651276
Update function arg names for plots
