+++ R/plot_slope.R
@@ -1,76 +1,83 @@
 #' Plot slope data for a 3d linestring with base R graphics
 #'
-#' @inheritParams slope_raster
-#' @inheritParams plot_dz
-#' @inheritParams sequential_dist
 #' @param route_xyz An sf linestring with x, y and z coordinates,
 #'   representing a route or other linear object.
+#' @param fill Should the profile be filled? `TRUE` by default
+#' @param horiz Should the legend be horizontal (`FALSE` by default)
+#' @param pal Color palette to use, `colorspace::diverging_hcl` by default.
+#' @param x Keyword, one of "bottomright", "bottom",
+#' "bottomleft", "left", "topleft", "top", "topright", "right" and "center"
+#' @param col Line colour, black by default
+#' @param cex Legend size, 0.9 by default
+#' @param bg Legend background colour, `grDevices::rgb(1, 1, 1, 0.8)` by default.
+#' @param title Title of the legend
+#' @param brks Breaks in colour palette to show.
+#'   `c(1, 3, 6, 10, 20, 40, 100)` by default.
+#' @param seq_brks Sequence of breaks to show in legend.
+#'   Includes negative numbers and omits zero by default
+#' @param ncol Number of columns in legend
+#' @param ... Additional parameters to pass to legend
+#' @inheritParams slope_raster
+#' @inheritParams sequential_dist
 #'
 #' @export
 #' @examples
 #' plot_slope(lisbon_route_3d)
 #' route_xyz = lisbon_road_segment_3d
 #' plot_slope(route_xyz)
-plot_slope = function(route_xyz,
-                      lonlat = sf::st_is_longlat(route_xyz),
-                      fill = TRUE,
-                      horiz = FALSE,
-                      p = colorspace::diverging_hcl,
-                      ...,
-                      x = "top",
-                      col = "black",
-                      cex = 0.9,
-                      bg = grDevices::rgb(1, 1, 1, 0.8),
-                      title = "Slope colors (percentage gradient)",
-                      s = 3:18,
-                      ncol = 4) {
+#' plot_slope(route_xyz, brks = c(1, 2, 4, 8, 16, 30))
+#' plot_slope(route_xyz, s = 5:8)
+plot_slope = function(
+  route_xyz,
+  lonlat = sf::st_is_longlat(route_xyz),
+  fill = TRUE,
+  horiz = FALSE,
+  pal =colorspace::diverging_hcl,
+  x = "top",
+  col = "black",
+  cex = 0.9,
+  bg = grDevices::rgb(1, 1, 1, 0.8),
+  title = "Slope colors (percentage gradient)",
+  brks = c(3, 6, 10, 20, 40, 100),
+  seq_brks = seq(from = 3, to = length(brks) * 2 - 2),
+  ncol = 4,
+  ...
+  ) {
   dz = distance_z(route_xyz, lonlat = lonlat)
-  plot_dz(dz$d, dz$z, fill = fill)
+  plot_dz(dz$d, dz$z, seq_brks = seq_brks, brks = brks, ...)
 }
 #' Plot a digital elevation profile based on xyz data
 #'
 #' @param d Cumulative distance
 #' @param z Elevations at points across a linestring
-#' @param p Color palette to use
-#' @param fill Should the profile be filled? `TRUE` by default
-#' @param x Keyword, one of "bottomright", "bottom",
-#' "bottomleft", "left", "topleft", "top", "topright", "right" and "center"
-#' @param col Line colour
-#' @param cex Legend size
-#' @param bg Legend background colour
[…]
