+++ R/slopes.R
@@ -57,6 +57,9 @@ slope_distance_weighted = function(d, elevations) {
 }
 #' Calculate the gradient of line segments from a 3D matrix of coordinates
 #'
+#' @param elevations Elevations in same units as x (assumed to be metres).
+#'   Default value: `m[, 3]`, meaning the 'z' coordinate in a matrix of
+#'   coordinates.
 #' @param m Matrix containing coordinates and elevations
 #' @inheritParams slope_vector
 #' @inheritParams sequential_dist
@@ -90,11 +93,11 @@ slope_matrix_weighted = function(m, elevations = m[, 3], lonlat = TRUE) {
 
 #' Calculate the sequential distances between sequential coordinate pairs
 #'
-#' @param lonlat Are the coordinates in lon/lat (geographic) coordinates?
-#'  Set this to `FALSE` if you have projected data, e.g. with coordinates
-#'  representing distance in meters, not degrees. Lonlat coodinates are assumed
-#'  (`lonlat = TRUE`) is the default. In `slope_raster()` and `plot_slope()`
-#'  it is taken from the CRS of the routes (`sf::st_is_longlat(routes)`).
+#' Set `lonlat` to `FALSE` if you have projected data, e.g. with coordinates
+#' representing distance in meters, not degrees. Lonlat coodinates are assumed
+#' (`lonlat = TRUE` is the default).
+#'
+#' @param lonlat Are the coordinates in lon/lat (geographic) coordinates? TRUE by default.
 #' @inheritParams slope_matrix
 #' @export
 #' @examples
@@ -130,7 +133,8 @@ slope_matrices = function(m_xyz_split, fun = slope_matrix_weighted, ...) {
 #' `r_linestring = sf::st_cast(routes, "LINESTRING")`.
 #'
 #' @inheritParams sequential_dist
-#' @inheritParams elevation_extract
+#' @param lonlat Are the routes provided in longitude/latitude coordinates?
+#'   By default, value is from the CRS of the routes (`sf::st_is_longlat(routes)`).
 #' @param routes Routes, the gradients of which are to be calculated.
 #'   The object must be of class `sf` with `LINESTRING` geometries.
 #' @param dem Raster overlapping with `routes` and values representing elevations
@@ -147,7 +151,7 @@ slope_matrices = function(m_xyz_split, fun = slope_matrix_weighted, ...) {
 #' (s = slope_raster(routes, dem))
 #' cor(routes$Avg_Slope, s)
 slope_raster = function(
- routes,
+  routes,
   dem = NULL,
   lonlat = sf::st_is_longlat(routes),
   method = "bilinear",
@@ -167,20 +171,21 @@ slope_raster = function(
 
 #' Extract slopes from xyz data frame or sf objects
 #'
-#' @param r_xyz An sf object with x, y, z dimensions
+#' @param route_xyz An sf object with x, y, z dimensions
+#' @param lonlat Are the coordinates in lon/lat order? TRUE by default
 #' @inheritParams slope_raster
 #'
 #' @export
 #' @examples
-#' r_xyz = lisbon_road_segment_3d
-#' slope_xyz(r_xyz, lonlat = FALSE)
-slope_xyz = function(r_xyz, fun = slope_matrix_weighted, lonlat = TRUE) {
-  if(inherits(r_xyz, "sf") | inherits(r_xyz, "sfc")) {
-    lonlat = sf::st_is_longlat(r_xyz)
-    r_xyz = as.data.frame(sf::st_coordinates(r_xyz))
+#' route_xyz = lisbon_road_segment_3d
[…]
