### 2021-08-17 f666563
Move terra param to slope_raster
