### 2021-08-17 018d2d6
usethis::use_package_doc() - package help
For #32
