+++ R/slopes-package.R
@@ -1,0 +1,27 @@
+#' @title slopes: Calculate Slopes of Roads, Rivers and Trajectories
+#'
+#' @description Calculate the slope (also known as steepness and gradient)
+#' of linear features such as roads, rivers and ski routes.
+#'
+#' Created by Robin Lovelace and Rosa Félix, who developed the package
+#' to support their research in active travel, and the prioritisation of
+#' investment in cycling, an activity which is sensitive to slopes,
+#' in particular.
+#'
+#' See [github.com/ITSLeeds/slopes](https://github.com/ITSLeeds/slopes/)
+#' for the source code and [itsleeds.github.io/slopes](https://itsleeds.github.io/slopes/)
+#' for the website.
+#'
+#' All package functions can be found in the
+#' [Reference page](https://itsleeds.github.io/slopes/reference/index.html)
+#' on the package website.
+#'
+
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL
