+++ man/plot_slope.Rd
@@ -10,7 +10,7 @@ plot_slope(
   fill = TRUE,
   horiz = FALSE,
   pal = colorspace::diverging_hcl,
-  x = "top",
+  legend_position = "top",
   col = "black",
   cex = 0.9,
   bg = grDevices::rgb(1, 1, 1, 0.8),
@@ -34,8 +34,9 @@ By default, value is from the CRS of the routes (\code{sf::st_is_longlat(routes)
 
 \item{pal}{Color palette to use, \code{colorspace::diverging_hcl} by default.}
 
-\item{x}{Keyword, one of "bottomright", "bottom",
-"bottomleft", "left", "topleft", "top", "topright", "right" and "center"}
+\item{legend_position}{The legend position. One of "bottomright", "bottom",
+"bottomleft", "left", "topleft", "top" (the default), "topright", "right"
+and "center".}
 
 \item{col}{Line colour, black by default}
 
@@ -43,7 +44,7 @@ By default, value is from the CRS of the routes (\code{sf::st_is_longlat(routes)
 
 \item{bg}{Legend background colour, \code{grDevices::rgb(1, 1, 1, 0.8)} by default.}
 
-\item{title}{Title of the legend}
+\item{title}{Title of the legend, \code{NULL} by default.}
 
 \item{brks}{Breaks in colour palette to show.
 \code{c(1, 3, 6, 10, 20, 40, 100)} by default.}
@@ -51,7 +52,7 @@ By default, value is from the CRS of the routes (\code{sf::st_is_longlat(routes)
 \item{seq_brks}{Sequence of breaks to show in legend.
 Includes negative numbers and omits zero by default}
 
-\item{ncol}{Number of columns in legend}
+\item{ncol}{Number of columns in legend, 4 by default.}
 
 \item{...}{Additional parameters to pass to legend}
 }
