### 2021-08-17 fe5f12d
Remove commented out examples for #32
