+++ data-raw/cyclestreets_route.R
@@ -1,0 +1,6 @@
+# Get data from cyclestreets
+cyclestreets_route = cyclestreets::journey(
+  stplanr::geo_code("Potternewton Park Leeds"),
+  stplanr::geo_code("University of Leeds")
+)
+usethis::use_data(cyclestreets_route)
