+++ R/data.R
@@ -43,16 +43,16 @@
 #' Road segments representing suggested route to cycle
 #' in Leeds, UK.
 #'
+#' Simple feature collection with 30 features and 32 fields
+#'
+#' See `data-raw` folder in the package's github repo for details.
+#'
 #' @format An object of class `sf`
 #' @source CycleStreets.net
 #' @examples
-#' # cyclestreets_route = cyclestreets::journey(
-#' #   stplanr::geo_code("Potternewton Park Leeds"),
-#' #   stplanr::geo_code("University of Leeds")
-#' # )
-#' # usethis::use_data(cyclestreets_route)
+#' library(sf)
 #' class(cyclestreets_route)
-#' # plot(cyclestreets_route)
+#' plot(cyclestreets_route$geometry)
 "cyclestreets_route"
 
 #' Road segments in Magnolia, Seattle
