+++ R/z.R
@@ -11,7 +11,7 @@
 #' sf::st_geometry(x)
 #' z_value(x)[1:5]
 #' xy = slopes::lisbon_route
-#' # z_value(xy) # error message
+#' try(z_value(xy)) # error message
 #' z_start(x)
 #' z_end(x)
 z_value = function(x) {
