+++ man/elevation_extract.Rd
@@ -32,10 +32,15 @@ m = sf::st_coordinates(lisbon_road_segments[1, ])
 dem = dem_lisbon_raster
 elevation_extract(m, dem)
 elevation_extract(m, dem, method = "simple")
-# uncomment the following lines to test with terra (experimental)
-# u = paste0("https://github.com/ITSLeeds/slopes/",
-#    "releases/download/0.0.0/dem_lisbon.tif" )
-# if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
-# et = terra::rast("dem_lisbon.tif")
-# elevation_extract(m, et)
+# Test with terra (requires internet connection):
+\donttest{
+if(has_terra()) {
+u = paste0("https://github.com/ITSLeeds/slopes/",
+   "releases/download/0.0.0/dem_lisbon.tif" )
+if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
+et = terra::rast("dem_lisbon.tif")
+elevation_extract(m, et)
+file.remove("dem_lisbon.tif")
+}
+}
 }
