+++ man/elevation_get.Rd
@@ -18,12 +18,13 @@ The object must be of class \code{sf} with \code{LINESTRING} geometries.}
 Get elevation data from hosted maptile services
 }
 \examples{
-# time-consuming, and see
-# https://github.com/ITSLeeds/slopes/runs/648128378#step:18:107
+# Time-consuming examples that require an internet connectioin and API key:
+\donttest{
 routes = cyclestreets_route
-# e = elevation_get(routes)
-# class(e)
-# e
-# plot(e)
-# plot(sf::st_geometry(routes), add = TRUE)
+e = elevation_get(routes)
+class(e)
+e
+plot(e)
+plot(sf::st_geometry(routes), add = TRUE)
+}
 }
