### 2021-08-17 d7869d0
Add clear error message with no CRS
Close #36 also for #32
