+++ R/plot_slope.R
@@ -44,6 +44,12 @@ plot_slope = function(
   ncol = 4,
   ...
   ) {
+  if(is.na(lonlat)) {
+    stop(
+      "CRS of routes not known. Set the CRS, e.g. as follows:\n",
+      "sf::st_crs(routes) = 4326 # if the routes are in lon/lat coordinates"
+    )
+  }
   dz = distance_z(route_xyz, lonlat = lonlat)
   plot_dz(dz$d, dz$z, seq_brks = seq_brks, brks = brks, ...)
 }
