### 2021-08-17 9d7bddd
Pass args from plot_slope() to plot_dz()
For #32
