+++ R/slope_get.R
@@ -7,6 +7,8 @@
 #' @examples
 #' # Time-consuming examples that require an internet connectioin and API key:
 #' \donttest{
+#' library(sf)
+#' library(raster)
 #' routes = cyclestreets_route
 #' e = elevation_get(routes)
 #' class(e)
