+++ R/slopes.R
@@ -215,7 +215,7 @@ slope_xyz = function(route_xyz, fun = slope_matrix_weighted, lonlat = TRUE) {
 #' elevation_extract(m, dem, method = "simple")
 #' # Test with terra (requires internet connection):
 #' \donttest{
-#' if(has_terra()) {
+#' if(slopes:::has_terra()) {
 #' u = paste0("https://github.com/ITSLeeds/slopes/",
 #'    "releases/download/0.0.0/dem_lisbon.tif" )
 #' if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
