### 2021-08-17 35adc64
Fix actions on window, mode = "wb"
