+++ man/elevation_extract.Rd
@@ -37,7 +37,8 @@ elevation_extract(m, dem, method = "simple")
 if(slopes:::has_terra()) {
 u = paste0("https://github.com/ITSLeeds/slopes/",
    "releases/download/0.0.0/dem_lisbon.tif" )
-if(!file.exists("dem_lisbon.tif")) download.file(u, "dem_lisbon.tif")
+f = "dem_lisbon.tif"
+if(!file.exists(f)) download.file(u, f, mode = "wb")
 et = terra::rast("dem_lisbon.tif")
 elevation_extract(m, et)
 file.remove("dem_lisbon.tif")
