### 2021-08-18 7ec8138
Suppress warnings, close #37
