+++ README.Rmd
@@ -21,7 +21,13 @@ knitr::opts_chunk$set(
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/420_status.svg)](https://github.com/ropensci/software-review/issues/420)
 <!-- badges: end -->
 
-This package calculates longitudinal steepness of linear features such as roads and rivers, based on two main inputs: vector linestring geometries and raster digital elevation model (DEM) datasets.
+The **slopes** R package calculates the slope (longitudinal steepness, also known as gradient) of linear features such as roads and rivers, based on two main inputs:
+
+- [vector](https://geocompr.robinlovelace.net/spatial-class.html#vector-data) linestring geometries defined by classes in the [`sf` package](https://r-spatial.github.io/sf/)
+- [raster](https://geocompr.robinlovelace.net/spatial-class.html#raster-data) objects with pixel values reporting average height, commonly known as digital elevation model (DEM) datasets, defined by classes in the [`raster`](https://cran.r-project.org/package=raster) or more recent [`terra`](https://rspatial.org/terra) packages
+
+This README covers installation and basic usage.
+For more information about slopes and how to use the package to calculate them, see the [slopes vignette](https://itsleeds.github.io/slopes/).
 
 ## Installation
 
@@ -53,7 +59,7 @@ usethis::edit_r_environ()
 MAPBOX_API_KEY=xxxxx
 ```
 
-## Usage
+## Basic example
 
 Load the package in the usual way:
 
@@ -67,8 +73,87 @@ We will also load the `sf` library:
 library(sf)
 ```
 
-The minimum data requirements for using the package are elevation points, either as a vector, a matrix or as a digital elevation model (DEM) encoded as a raster dataset.
-Typically you will also have a geographic object representing the roads or similar features.
+The minimum input data requirement for using the package is an `sf` object containing LINESTRING geometries.
+You can check your input dataset is suitable with the functions `class()` from base R and `st_geometry_type()` from the `sf` package, as demonstrated below on the object `lisbon_road_segment` that is contained within the package:
+
+```{r}
+class(lisbon_road_segment)
+st_geometry_type(lisbon_road_segment)
+```
+
+Don't worry if you don't yet have your linear features in this class: you can read-in data from a wide range of formats into an `sf` object.
+You can also create `sf` objects from a matrix of coordinates, as illustrated below (don't worry about the details for now, you can read up on how all this works in the `sf` package [documentation](https://r-spatial.github.io/sf/articles/sf1.html)):
+
+```{r, eval=FALSE, echo=FALSE}
+m = st_coordinates(sf::st_transform(lisbon_road_segment, 4326))
+s = seq(from = 1, to = nrow(m), length.out = 4)
+round(m[s, 1:2], 5)
+dput(round(m[s, 1], 4))
+dput(round(m[s, 2], 4))
+```
+
+```{r}
+m = cbind(
+  c(-9.1333, -9.134, -9.13),
+  c(38.714, 38.712, 38.710)
+)
+sf_linestring = sf::st_sf(
+  data.frame(id = 1),
+  geometry = st_sfc(st_linestring(m)),
+  crs = 4326
+)
+class(sf_linestring)
+st_geometry_type(sf_linestring)
+```
+
+A quick way of testing if your object can have slopes calculated for it is to plot it in an interactive map and to check that underneath the object there is indeed terrain that will give the linestrings gradient:
+
+```{r linestringmap}
+library(tmap)
+tmap_mode("view")
+tm_shape(sf_linestring) +
+  tm_lines(lwd = 5) +
[…]
