### 2021-08-18 3d324fe
Improve docs for elevation_get()
