### 2021-08-18 #420 @Robinlovelace
Update on this, we're making progress! We've addressed around half of the issues and think the others are doable. See here for updates (23 down, 23 to go at the time of writing): https://github.com/ITSLeeds/slopes/issues/32
