+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# slopes
+# slopes package
 
 <!-- badges: start -->
 
@@ -13,23 +13,43 @@ Review](https://badges.ropensci.org/420_status.svg)](https://github.com/ropensci
 <!-- badges: end -->
 
 The **slopes** R package calculates the slope (longitudinal steepness,
-also known as gradient) of linear features such as roads and rivers,
-based on two main inputs:
+also known as gradient) of roads, rivers and other linear (simple)
+features, based on two main inputs:
 
 -   [vector](https://geocompr.robinlovelace.net/spatial-class.html#vector-data)
-    linestring geometries defined by classes in the [`sf`
-    package](https://r-spatial.github.io/sf/)
+    linestring geometries defined by classes in the
+    [`sf`](https://r-spatial.github.io/sf/) package
 -   [raster](https://geocompr.robinlovelace.net/spatial-class.html#raster-data)
     objects with pixel values reporting average height, commonly known
-    as digital elevation model (DEM) datasets, defined by classes in the
-    [`raster`](https://cran.r-project.org/package=raster) or more recent
-    [`terra`](https://rspatial.org/terra) packages
+    as digital elevation model (**DEM**) datasets, defined by classes in
+    the [`raster`](https://cran.r-project.org/package=raster) or more
+    recent [`terra`](https://rspatial.org/terra) packages
 
 This README covers installation and basic usage. For more information
-about slopes and how to use the package to calculate them, see the
-[slopes vignette](https://itsleeds.github.io/slopes/).
+about slopes and how to use the package to calculate them, see the [get
+started](https://itsleeds.github.io/slopes/) and the [introducion to
+slopes](https://itsleeds.github.io/intro-to-slopes/) vignette.
 
-## Installation
+## How it works
+
+The package takes two main types of input data for slope calculation: -
+vector geographic objects representing **linear features**, and -
+**elevation values** from a digital elevation model representing a
+continuous terrain surface or which can be downloaded using
+functionality in the package
+
+The package can be used with two sources of elevation data: - openly
+available elevation data via an interface to the [ceramic
+package](https://github.com/hypertidy/ceramic), enabling estimation of
+hilliness for routes anywhere worldwide even when local elevation data
+is lacking. The package takes geographic lines objects and returns
+elevation data per vertex (providing the output as a 3D point geometry
+in the sf package by default) and per line feature (providing average
+gradient by default). - an elevation model, available on your machine.
+
+## Getting started
+
+### Installation
 
 <!-- You can install the released version of slopes from [CRAN](https://CRAN.R-project.org) with: -->
 <!-- ``` r -->
@@ -43,11 +63,11 @@ Install the development version from [GitHub](https://github.com/) with:
 remotes::install_github("itsleeds/slopes")
 ```
 
-### Installation for DEM downloads
+#### Installation for DEM downloads
 
 If you do not already have DEM data and want to make use of the
 package’s ability to download them using the `ceramic` package, install
[…]
