### 2021-08-24 e2122e1
move "slopes" vignette to "intro to slopes", and add information.
