+++ vignettes/slopes.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "An introduction to the slopes package"
+title: "Get started"
 output: bookdown::html_vignette2
 vignette: >
-  %\VignetteIndexEntry{An introduction to the slopes package}
+  %\VignetteIndexEntry{Get started}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 bibliography: slope-references.bib
@@ -15,103 +15,100 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Introduction
+## Installation
+(again)
 
-Slopes are important for many purposes, including flood risk, agriculture, geology, and infrastructure constructions. The information about slopes is very relevant for vehicle routing modes that are sensitive to going up or down steep hills (e.g. bicycles, trains, and large trucks).  
-![Traffic sign](https://sinalnorte.com/wp-content/uploads/2018/02/A3b.jpg){ width=10% }  
+## Functions 
 
-Although there are several ways to name "slope", such as "steepness", "hilliness", "inclination", "aspect", "gradient", "declivity", the referred `slopes` in this package can be defined as the "longitudinal gradient" [@cohen_global_2018].
+### Elevation
 
-# Calculating slopes
+elevation_add() Take a linestring and add a third dimension (z) to its coordinates
+elevation_get() Get elevation data from hosted maptile services
+elevation_extract() Extract elevations from coordinates
 
-The most common slope calculation method is defined by the vertical difference of the final and start point or line height (z1 and z0) divided by the horizontal length that separates them.
+z_value() retrieves elevation values for each z (as vector of sequential vertices)
+z_start() retrieves the elevation value of the first linestring vertice
+z_end() retrieves the elevation value of the last linestring vertice
+z_mean() retrieves the elevation mean value 
+z_max()  retrieves the elevation max value 
+z_min()  retrieves the elevation min value 
 
-$$
-s = \Delta z/l
-$$
+### Distance
 
-Depending on the purpose of application, it might me relevant to understand how hilliness is estimated.  
+sequential_dist() Calculate the sequential distances between sequential coordinate pairs
 
-# Implementation in the slopes package
+### Slope
 
-Using the `slopes` package we can estimate the gradient of individual road segments. When these segments are combined into routes, we then need a means of assessing the hilliness of the entire route. A range of indices can be used to represent route hilliness. The choice of which index is most appropriate may be context dependent.
+slope_vector() calculates the slopes associated with consecutive elements in one dimensional distance and associated elevations.
+slope_distance() calculates the slopes associated with consecutive distances and elevations.
+slope_distance_mean() calculates the mean average slopes associated with consecutive distances and elevations.
+slope_distance_weighted() calculates the slopes associated with consecutive distances and elevations, with the mean value associated with each set of distance/elevation vectors weighted in proportion to the distance between each elevation measurement, so longer sections have proportionally more influence on the resulting gradient estimate.
 
-To help research slopes using open source software for reproducibility, we have developed the slopes package.
-After you have installed it, load it as follows:
+slope_raster() Calculate slopes of linestrings based on local raster map
+slope_matrix() Calculate the gradient of line segments from a 3D matrix of coordinates
+slope_matrix_weighted() Calculate the weighted gradient of line segments from a 3D matrix of coordinates
 
-```{r setup}
-library(slopes)
-```
[…]
