+++ vignettes/slopes.Rmd
@@ -16,67 +16,373 @@ knitr::opts_chunk$set(
 ```
 
 ## Installation
-(again)
+
+<!-- You can install the released version of slopes from [CRAN](https://CRAN.R-project.org) with: -->
+
+<!-- ``` r -->
+<!-- install.packages("slopes") -->
+<!-- ``` -->
+
+Install the development version from [GitHub](https://github.com/) with:
+
+```{r, eval=FALSE}
+# install.packages("remotes")
+remotes::install_github("itsleeds/slopes")
+```
+
+#### Installation for DEM downloads
+
+If you do not already have DEM data and want to make use of the package's ability to download them using the `ceramic` package, install the package with suggested dependencies, as follows:
+
+```{r, eval=FALSE}
+# install.packages("remotes")
+remotes::install_github("itsleeds/slopes", dependencies = "Suggests")
+```
+
+Furthermore, you will need to add a MapBox API key to be able to get DEM datasets, by signing up and registering for a key at https://account.mapbox.com/access-tokens/ and then following these steps:
+
+```{r, eval=FALSE}
+usethis::edit_r_environ()
+MAPBOX_API_KEY=xxxxx # replace XXX with your api key
+```
+
 
 ## Functions 
 
 ### Elevation
 
-elevation_add() Take a linestring and add a third dimension (z) to its coordinates
-elevation_get() Get elevation data from hosted maptile services
-elevation_extract() Extract elevations from coordinates
+- `elevation_add()` Take a linestring and add a third dimension (z) to its coordinates
+- `elevation_get()` Get elevation data from hosted maptile services
+- `elevation_extract()` Extract elevations from coordinates
 
-z_value() retrieves elevation values for each z (as vector of sequential vertices)
-z_start() retrieves the elevation value of the first linestring vertice
-z_end() retrieves the elevation value of the last linestring vertice
-z_mean() retrieves the elevation mean value 
-z_max()  retrieves the elevation max value 
-z_min()  retrieves the elevation min value 
+- `z_value()` retrieves elevation values for each z (as vector of sequential vertices)
+- `z_start()` retrieves the elevation value of the first linestring vertice
+- `z_end()` retrieves the elevation value of the last linestring vertice
+- `z_mean()` retrieves the elevation mean value 
+- `z_max()`  retrieves the elevation max value 
+- `z_min()`  retrieves the elevation min value 
 
 ### Distance
 
-sequential_dist() Calculate the sequential distances between sequential coordinate pairs
+- `sequential_dist()` Calculate the sequential distances between sequential coordinate pairs
 
 ### Slope
 
-slope_vector() calculates the slopes associated with consecutive elements in one dimensional distance and associated elevations.
-slope_distance() calculates the slopes associated with consecutive distances and elevations.
-slope_distance_mean() calculates the mean average slopes associated with consecutive distances and elevations.
-slope_distance_weighted() calculates the slopes associated with consecutive distances and elevations, with the mean value associated with each set of distance/elevation vectors weighted in proportion to the distance between each elevation measurement, so longer sections have proportionally more influence on the resulting gradient estimate.
+- `slope_vector()` calculates the slopes associated with consecutive elements in one dimensional distance and associated elevations.
[…]
