### 2021-08-24 c424e9a
remove bibliography in get started.
