### 2021-08-27 6803fcc
Do not download mapbox tiles in vignette
try to fix build!
