+++ man/lisbon_road_segment.Rd
@@ -20,8 +20,11 @@ lisbon_road_segment
 }
 \description{
 A single road segment and a 3d version.
+Different versions of this dataset are provided.
 }
 \details{
+The \code{lisbon_road_segment} has ... columns and ... rows.
+
 The \code{lisbon_road_segment_xyz_mapbox} was created with:
 \code{lisbon_road_segment_xyz_mapbox = elevation_add(lisbon_road_segment)}.
 }
