+++ man/dem_lisbon_raster.Rd
@@ -18,6 +18,14 @@ dem_lisbon_raster
 A dataset containing elevation in and around Lisbon
 with a geographic resolution of 10m.
 }
+\details{
+The dataset orginates from This DEM was acquired by Instituto Superior
+Técnico (University of Lisbon) by 2012, covers all the Northern
+Metropolitan Area of Lisbon, and has a 10m cell resolution,
+when projected at the official Portuguese EPSG: 3763 - TM06/ETRS89.
+The dataset was released as an open access dataset with permission from the
+University of Lisbon to support this project.
+}
 \examples{
 names(dem_lisbon_raster)
 raster::plot(dem_lisbon_raster)
