+++ man/lisbon_route.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{lisbon_route}
+\alias{lisbon_route}
+\alias{lisbon_route_3d}
+\alias{lisbon_route_xyz_mapbox}
+\title{A route composed of a single linestring in Lisbon, Portugal}
+\format{
+An object of class \code{sf}
+}
+\source{
+Produced by ESRI's
+\href{https://pro.arcgis.com/en/pro-app/help/analysis/}{3D Analyst extension}
+}
+\usage{
+lisbon_route
+}
+\description{
+A route representing a trip from the Santa Catarina area
+in the East of central Lisbon the map to the Castelo de São Jorge
+in the West of central Lisbon.
+}
+\details{
+Different versions of this dataset are provided.
+
+The \code{lisbon_route} object has 1 row and 4 columns.
+
+The \code{lisbon_route_xyz_mapbox} was created with:
+\code{lisbon_route_xyz_mapbox = elevation_add(lisbon_route)}.
+}
+\examples{
+lisbon_route
+lisbon_route_3d
+lisbon_route_xyz_mapbox
+}
+\keyword{datasets}
