+++ vignettes/slopes.Rmd
@@ -279,7 +279,6 @@ slope_distance_mean(dist, elevations)
 slope_distance_weighted(dist, elevations)
 ```
 
-
 In any case, to use the `slopes` package you need **elevation points**, either as a vector, a matrix or as a digital elevation model (DEM) encoded as a raster dataset.
 
 ## Calculating and plotting the gradient of roads
@@ -325,7 +324,8 @@ raster::plot(dem_lisbon_raster)
 plot(lisbon_road_network["slope"], add = TRUE, lwd = 5)
 ```
 
-Imagine that we want to go from East hilly area of the map (Santa Catarina) to the West hilly area of the map (Castelo de São Jorge):
+Imagine that we want to go from from the Santa Catarina area in the East of the map to the Castelo de São Jorge in the West.
+This route goes down a valley and up the other side:
 
 ```{r route}
 # library(tmap)
@@ -394,10 +394,6 @@ cyclestreets_xyz = elevation_add(cyclestreets_route)
 plot_slope(cyclestreets_xyz)
 plot(cumsum(cyclestreets_xyz$distances), cumsum(cyclestreets_xyz$elevation_change))
 ```
-
-
-
-
 
 
 ## See more in vignettes
