### 2021-08-27 afb508d
Update vignette, fix pkgdown website
