+++ vignettes/roadnetworkcycling.Rmd
@@ -193,8 +193,8 @@ st_write(network, "shapefiles/SlopesIoW.gpkg", append=F)
 -   [São Paulo (Brazil)](http://web.tecnico.ulisboa.pt/~rosamfelix/gis/declives/DeclivesSaoPaulo.html)
 -   [Lisbon (Portugal)](http://web.tecnico.ulisboa.pt/~rosamfelix/gis/declives/DeclivesLisboa.html)
 -   [Oporto (Portugal)](http://web.tecnico.ulisboa.pt/~rosamfelix/gis/declives/DeclivesPorto_EU.html)
+-   [Leeds (UK)](http://web.tecnico.ulisboa.pt/~rosamfelix/gis/declives/SlopesLeeds.html)
 -   Bristol (UK)
--   Sheffield (UK)
 
 
 ```{r tidyup, include=FALSE}
