### 2021-08-30 d903043
Fix preparation for pkgdown & README redux
It will not run ok because we need to add the `lisbon_road_segment_xyz_mapbox` and  `lisbon_route_xyz_mapbox` as available datasets!
#40
