+++ README.Rmd
@@ -85,43 +85,15 @@ The minimum input data requirement for using the package is an `sf` object conta
 
 ### Sinlge road segment + no DEM
 
-You can check your input dataset is suitable with the functions `class()` from base R and `st_geometry_type()` from the `sf` package, as demonstrated below on the example object `lisbon_road_segment` that is contained within the package:
+In this example, we use a single linestring object `lisbon_road_segment` that is contained within the package.
 
 ```{r}
-sf_linestring = lisbon_road_segment
-class(sf_linestring)
-st_geometry_type(sf_linestring)
+sf_linestring = lisbon_road_segment # import or load a linestring
 ```
-
-Don't worry if you don't yet have your linear features in this class: you can read-in data from a wide range of formats into an `sf` object.
-
-<!-- You can also create `sf` objects from a matrix of coordinates, as illustrated below (don't worry about the details for now, you can read up on how all this works in the `sf` package [documentation](https://r-spatial.github.io/sf/articles/sf1.html)): -->
-
-<!-- ```{r, eval=FALSE, echo=FALSE} -->
-<!-- m = st_coordinates(sf::st_transform(lisbon_road_segment, 4326)) -->
-<!-- s = seq(from = 1, to = nrow(m), length.out = 4) -->
-<!-- round(m[s, 1:2], 5) -->
-<!-- dput(round(m[s, 1], 4)) -->
-<!-- dput(round(m[s, 2], 4)) -->
-<!-- ``` -->
-
-<!-- ```{r} -->
-<!-- m = cbind( -->
-<!--   c(-9.1333, -9.134, -9.13), -->
-<!--   c(38.714, 38.712, 38.710) -->
-<!-- ) -->
-<!-- sf_linestring = sf::st_sf( -->
-<!--   data.frame(id = 1), -->
-<!--   geometry = st_sfc(st_linestring(m)), -->
-<!--   crs = 4326 -->
-<!-- ) -->
-<!-- class(sf_linestring) -->
-<!-- st_geometry_type(sf_linestring) -->
-<!-- ``` -->
 
 A quick way of testing if your object can have slopes calculated for it is to plot it in an interactive map and to check that underneath the object there is indeed terrain that will give the linestrings gradient:
 
-```{r linestringmap, message=FALSE, warning=FALSE}
+```{r}
 library(tmap)
 tmap_mode("view")
 tm_shape(sf_linestring) +
@@ -129,22 +101,23 @@ tm_shape(sf_linestring) +
   tm_basemap(leaflet::providers$Esri.WorldTopoMap)
 ```
 
-Imagine you want to calculate the gradient of the route shown above. 
-You can do this as a two step process as follows.
+we can calculate its gradient as a two step process, as follows.
 
 **Step 1**: add elevations to each coordinate in the linestring (requires a [MapBox API](https://account.mapbox.com/access-tokens/) key):
 
-```{r}
+```{r, eval=FALSE}
 sf_linestring_xyz = elevation_add(sf_linestring)
+#> Loading required namespace: ceramic
+#> Preparing to download: 9 tiles at zoom = 18 from 
+#> https://api.mapbox.com/v4/mapbox.terrain-rgb/
+```
+```{r, echo=FALSE}
+# note: the following should be TRUE
+# identical(sf_linestring_xyz, lisbon_road_segment_xyz_mapbox)
+sf_linestring_xyz = lisbon_road_segment_xyz_mapbox
 ```
 
-You can check the elevations added to the new `sf_linestring_xyz` object by printing its coordinates, as follows (note the new Z column that goes from above 87 m above sea level to only 79 m in a short distance).
-
-```{r}
-st_coordinates(sf_linestring_xyz)
-```
-
-You can use the `z_` functions to extract such values:
[…]
