### 2021-08-30 d9c40ca
clarify dem= argument omit
in README and Get Started.
#32
