### 2021-08-30 52ad981
Shorten README
