+++ README.md
@@ -94,351 +94,45 @@ library(sf)
 ```
 
 The minimum input data requirement for using the package is an `sf`
-object containing LINESTRING geometries.
-
-### Sinlge road segment + no DEM
-
-You can check your input dataset is suitable with the functions
-`class()` from base R and `st_geometry_type()` from the `sf` package, as
-demonstrated below on the example object `lisbon_road_segment` that is
-contained within the package:
+object containing LINESTRING geometries, as illustrated below (requires
+a MapBox API key):
 
 ``` r
-sf_linestring = lisbon_road_segment
-class(sf_linestring)
-#> [1] "sf"         "tbl_df"     "tbl"        "data.frame"
-st_geometry_type(sf_linestring)
-#> [1] LINESTRING
-#> 18 Levels: GEOMETRY POINT LINESTRING POLYGON MULTIPOINT ... TRIANGLE
+sf_linestring = lisbon_route # import or load a linestring object
 ```
 
-Don’t worry if you don’t yet have your linear features in this class:
-you can read-in data from a wide range of formats into an `sf` object.
-
-<!-- You can also create `sf` objects from a matrix of coordinates, as illustrated below (don't worry about the details for now, you can read up on how all this works in the `sf` package [documentation](https://r-spatial.github.io/sf/articles/sf1.html)): -->
-<!-- ```{r, eval=FALSE, echo=FALSE} -->
-<!-- m = st_coordinates(sf::st_transform(lisbon_road_segment, 4326)) -->
-<!-- s = seq(from = 1, to = nrow(m), length.out = 4) -->
-<!-- round(m[s, 1:2], 5) -->
-<!-- dput(round(m[s, 1], 4)) -->
-<!-- dput(round(m[s, 2], 4)) -->
-<!-- ``` -->
-<!-- ```{r} -->
-<!-- m = cbind( -->
-<!--   c(-9.1333, -9.134, -9.13), -->
-<!--   c(38.714, 38.712, 38.710) -->
-<!-- ) -->
-<!-- sf_linestring = sf::st_sf( -->
-<!--   data.frame(id = 1), -->
-<!--   geometry = st_sfc(st_linestring(m)), -->
-<!--   crs = 4326 -->
-<!-- ) -->
-<!-- class(sf_linestring) -->
-<!-- st_geometry_type(sf_linestring) -->
-<!-- ``` -->
-
-A quick way of testing if your object can have slopes calculated for it
-is to plot it in an interactive map and to check that underneath the
-object there is indeed terrain that will give the linestrings gradient:
-
 ``` r
-library(tmap)
-tmap_mode("view")
-tm_shape(sf_linestring) +
-  tm_lines(lwd = 5) +
-  tm_basemap(leaflet::providers$Esri.WorldTopoMap)
-```
-
-<img src="man/figures/README-linestringmap-1.png" width="100%" />
-
-Imagine you want to calculate the gradient of the route shown above. You
-can do this as a two step process as follows.
-
-**Step 1**: add elevations to each coordinate in the linestring
-(requires a [MapBox API](https://account.mapbox.com/access-tokens/)
-key):
-
-``` r
-sf_linestring_xyz = elevation_add(sf_linestring)
+sf_linestring_xyz = elevation_add(sf_linestring)  # dem = NULL
 #> Loading required namespace: ceramic
-#> Preparing to download: 9 tiles at zoom = 18 from 
+#> Preparing to download: 12 tiles at zoom = 12 from 
 #> https://api.mapbox.com/v4/mapbox.terrain-rgb/
 ```
 
-You can check the elevations added to the new `sf_linestring_xyz` object
[…]
