+++ README.Rmd
@@ -81,248 +81,40 @@ library(slopes)
 library(sf)
 ```
 
-The minimum input data requirement for using the package is an `sf` object containing LINESTRING geometries.  
-
-### Sinlge road segment + no DEM
-
-In this example, we use a single linestring object `lisbon_road_segment` that is contained within the package.
+The minimum input data requirement for using the package is an `sf` object containing LINESTRING geometries, as illustrated below (requires a MapBox API key):
 
 ```{r}
-sf_linestring = lisbon_road_segment # import or load a linestring
+sf_linestring = lisbon_route # import or load a linestring object
 ```
-
-A quick way of testing if your object can have slopes calculated for it is to plot it in an interactive map and to check that underneath the object there is indeed terrain that will give the linestrings gradient:
-
-```{r}
-library(tmap)
-tmap_mode("view")
-tm_shape(sf_linestring) +
-  tm_lines(lwd = 5) +
-  tm_basemap(leaflet::providers$Esri.WorldTopoMap)
-```
-
-we can calculate its gradient as a two step process, as follows.
-
-**Step 1**: add elevations to each coordinate in the linestring (requires a [MapBox API](https://account.mapbox.com/access-tokens/) key):
 
 ```{r, eval=FALSE}
 sf_linestring_xyz = elevation_add(sf_linestring)  # dem = NULL
 #> Loading required namespace: ceramic
-#> Preparing to download: 9 tiles at zoom = 18 from 
-#> https://api.mapbox.com/v4/mapbox.terrain-rgb/
-```
-```{r, echo=FALSE}
-# note: the following should be TRUE
-# identical(sf_linestring_xyz, lisbon_road_segment_xyz_mapbox)
-sf_linestring_xyz = lisbon_road_segment_xyz_mapbox
-```
-
-With the argument `dem = NULL`, the function downloads the necessary elevation information from Mapbox. You can use this argument with a local digital elevation model (`dem = ...`).
-
-You can check the elevations added to the new `sf_linestring_xyz` object by printing its coordinates with `st_coordinates()`, or you can use the `z_` functions to extract such values, as follows:
-
-```{r}
-z_value(sf_linestring_xyz) # returns all the elevation values between xy coordinates
-
-z_mean(sf_linestring_xyz) # elevation mean value
-z_min(sf_linestring_xyz) # elevation min value 
-z_max(sf_linestring_xyz) # elevation max value 
-z_start(sf_linestring_xyz) # first z
-z_end(sf_linestring_xyz) # last z
-```
-
-Note the linestring goes from 88 m above sea level to only 80 m in a short distance.
-
-**Step 2**: calculate the average slope of the linestring
-
-```{r}
-slope_xyz(sf_linestring_xyz)
-```
-
-The result, just over 0.2, tells us that it's quite a steep slope: a 21% gradient *on average*.
-
-### Route + available DEM
-
-Using the slopes package we can estimate the gradient of individual road segments. When these segments are combined into routes, we then need a means of assessing the hilliness of the entire route. A range of indices can be used to represent route hilliness. The choice of which index is most appropriate may be context dependent (see the [introducion to slopes](https://itsleeds.github.io/intro-to-slopes/) vignette).
-  
-This example uses the same function with a entire route, `lisbon_route`, also available in the package.
-
-```{r message=FALSE, warning=FALSE}
-sf_route = lisbon_route
-
-tm_shape(sf_route) +
-  tm_lines(lwd = 3) +
[…]
