+++ pkgdown.yml
@@ -7,5 +7,5 @@ articles:
   roadnetworkcycling: roadnetworkcycling.html
   slopes: slopes.html
   verification: verification.html
-last_built: 2021-08-30T01:12Z
+last_built: 2021-08-30T19:35Z
 
