### 2021-08-30 7a8cdc5
Add @return values for #32
