+++ R/slopes.R
@@ -21,6 +21,15 @@
 #' @param x Vector of locations
 #' @param d Vector of distances between points
 #' @param elevations Elevations in same units as x (assumed to be metres)
+#' @return A vector of slope gradients associated with each linear element
+#'   (each line between consecutive vertices) associated with linear features.
+#'   Returned values for `slope_distance_mean()` and
+#'   `slope_distance_mean_weighted()` are summary statistics for all
+#'   linear elements in the linestring.
+#'   The output value is a proportion representing the change in elevation
+#'   for a given change in horizontal movement along the linestring.
+#'   0.02, for example, represents a low gradient of 2% while 0.08 represents
+#'   a steep gradient of 8%.
 #' @export
 #' @examples
 #' x = c(0, 2, 3, 4, 5, 9)
@@ -63,12 +72,23 @@ slope_distance_weighted = function(d, elevations) {
 #' @param m Matrix containing coordinates and elevations
 #' @inheritParams slope_vector
 #' @inheritParams sequential_dist
+#' @return A vector of slope gradients associated with each linear element
+#'   (each line between consecutive vertices) associated with linear features.
+#'   Returned values for `slope_matrix_mean()` and
+#'   `slope_matrix_weighted()` are summary statistics for all
+#'   linear elements in the linestring.
+#'   The output value is a proportion representing the change in elevation
+#'   for a given change in horizontal movement along the linestring.
+#'   0.02, for example, represents a low gradient of 2% while 0.08 represents
+#'   a steep gradient of 8%.
 #' @export
 #' @examples
 #' x = c(0, 2, 3, 4, 5, 9)
 #' y = c(0, 0, 0, 0, 0, 9)
 #' z = c(1, 2, 2, 4, 3, 1) / 10
 #' m = cbind(x, y, z)
+#' slope_matrix_weighted(m, lonlat = FALSE)
+#' slope_matrix_mean(m, lonlat = FALSE)
 #' plot(x, z, ylim = c(-0.5, 0.5), type = "l")
 #' (gx = slope_vector(x, z))
 #' (gxy = slope_matrix(m, lonlat = FALSE))
@@ -82,6 +102,13 @@ slope_matrix = function(m, elevations = m[, 3], lonlat = TRUE) {
   e_change = diff(elevations)
   g = e_change / d
   g
+}
+#' @rdname slope_matrix
+#' @export
+slope_matrix_mean = function(m, elevations = m[, 3], lonlat = TRUE) {
+  g1 = slope_matrix(m, elevations = elevations, lonlat = lonlat)
+  d = sequential_dist(m = m, lonlat = lonlat)
+  mean(abs(g1), na.rm = TRUE)
 }
 #' @rdname slope_matrix
 #' @export
@@ -99,12 +126,18 @@ slope_matrix_weighted = function(m, elevations = m[, 3], lonlat = TRUE) {
 #'
 #' @param lonlat Are the coordinates in lon/lat (geographic) coordinates? TRUE by default.
 #' @inheritParams slope_matrix
+#' @return A vector of distance values in meters if `lonlat = TRUE`
+#'   or the map units of the input data if `lonlat = FALSE` between
+#'   consecutive vertices.
 #' @export
 #' @examples
 #' x = c(0, 2, 3, 4, 5, 9)
 #' y = c(0, 0, 0, 0, 0, 1)
 #' m = cbind(x, y)
-#' sequential_dist(m)
+#' d = sequential_dist(m, lonlat = FALSE)
+#' d
+#' nrow(m)
+#' length(d)
 sequential_dist = function(m, lonlat = TRUE) {
   if(lonlat) {
[…]
