+++ man/slope_matrix.Rd
@@ -2,10 +2,13 @@
 % Please edit documentation in R/slopes.R
 \name{slope_matrix}
 \alias{slope_matrix}
+\alias{slope_matrix_mean}
 \alias{slope_matrix_weighted}
 \title{Calculate the gradient of line segments from a 3D matrix of coordinates}
 \usage{
 slope_matrix(m, elevations = m[, 3], lonlat = TRUE)
+
+slope_matrix_mean(m, elevations = m[, 3], lonlat = TRUE)
 
 slope_matrix_weighted(m, elevations = m[, 3], lonlat = TRUE)
 }
@@ -18,6 +21,17 @@ coordinates.}
 
 \item{lonlat}{Are the coordinates in lon/lat (geographic) coordinates? TRUE by default.}
 }
+\value{
+A vector of slope gradients associated with each linear element
+(each line between consecutive vertices) associated with linear features.
+Returned values for \code{slope_matrix_mean()} and
+\code{slope_matrix_weighted()} are summary statistics for all
+linear elements in the linestring.
+The output value is a proportion representing the change in elevation
+for a given change in horizontal movement along the linestring.
+0.02, for example, represents a low gradient of 2\% while 0.08 represents
+a steep gradient of 8\%.
+}
 \description{
 Calculate the gradient of line segments from a 3D matrix of coordinates
 }
@@ -26,6 +40,8 @@ x = c(0, 2, 3, 4, 5, 9)
 y = c(0, 0, 0, 0, 0, 9)
 z = c(1, 2, 2, 4, 3, 1) / 10
 m = cbind(x, y, z)
+slope_matrix_weighted(m, lonlat = FALSE)
+slope_matrix_mean(m, lonlat = FALSE)
 plot(x, z, ylim = c(-0.5, 0.5), type = "l")
 (gx = slope_vector(x, z))
 (gxy = slope_matrix(m, lonlat = FALSE))
