### 2021-08-30 7a67e2d
Update elevation_get() docs for #32
