### 2021-08-30 18b0a5d
Improve documentation of elevation_extract() for #32
