+++ man/elevation_extract.Rd
@@ -28,7 +28,30 @@ if \code{dem} is of class \code{SpatRast}}
 A vector of elevation values.
 }
 \description{
-Extract elevations from coordinates
+This function takes a series of points located in geographical space
+and a digital elevation model as inputs and returns a vector of
+elevation estimates associated with each point.
+The function takes locations
+represented as a matrix of XY (or longitude latitude) coordinates
+and a digital elevation model (DEM) with class \code{raster} or \code{terra}.
+It returns a vector of values representing estimates of elevation
+associated with each of the points.
+}
+\details{
+By default, the elevations are estimated using
+\href{https://en.wikipedia.org/wiki/Bilinear_interpolation}{bilinear interpolation}
+(\code{method = "bilinear"})
+which calculates point height based on proximity to the centroids of
+surrounging cells.
+The value of the \code{method} argument is passed to the \code{method} argument in
+\href{https://rspatial.github.io/raster/reference/extract.html}{\code{raster::extract()}}
+or
+\href{https://rspatial.github.io/terra/reference/extract.html}{\code{terra::extract()}}
+depending on the class of the input raster dataset.
+
+See Kidner et al. (1999)
+for descriptions of alternative elevation interpolation and extrapolation
+algorithms.
 }
 \examples{
 m = sf::st_coordinates(lisbon_road_network[1, ])
@@ -47,4 +70,10 @@ elevation_extract(m, et)
 file.remove("dem_lisbon.tif")
 }
 }
+}
+\references{
+Kidner, David, Mark Dorey, and Derek Smith.
+"What’s the point? Interpolation and extrapolation with a regular grid DEM."
+Fourth International Conference on GeoComputation, Fredericksburg,
+VA, USA. 1999.
 }
