### 2021-08-30 7637f45
Improve docs for sequential_dist(), see #32
