+++ R/slopes.R
@@ -69,7 +69,13 @@ slope_distance_weighted = function(d, elevations) {
 #' @param elevations Elevations in same units as x (assumed to be metres).
 #'   Default value: `m[, 3]`, meaning the 'z' coordinate in a matrix of
 #'   coordinates.
-#' @param m Matrix containing coordinates and elevations
+#' @param m Matrix containing coordinates and elevations.
+#'   The matrix should have three columns: x, y, and z. Typically
+#'   these correspond to location in the West-East, South-North, and vertical
+#'   elevation axes respectively.
+#'   In data with geographic coordinates, Z values are assumed to be in
+#'   metres. In data with projected coordinates, Z values are assumed to have
+#'   the same units as the X and Y coordinates.
 #' @inheritParams slope_vector
 #' @inheritParams sequential_dist
 #' @return A vector of slope gradients associated with each linear element
