+++ man/elevation_extract.Rd
@@ -12,7 +12,13 @@ elevation_extract(
 )
 }
 \arguments{
-\item{m}{Matrix containing coordinates and elevations}
+\item{m}{Matrix containing coordinates and elevations.
+The matrix should have three columns: x, y, and z. Typically
+these correspond to location in the West-East, South-North, and vertical
+elevation axes respectively.
+In data with geographic coordinates, Z values are assumed to be in
+metres. In data with projected coordinates, Z values are assumed to have
+the same units as the X and Y coordinates.}
 
 \item{dem}{Raster overlapping with \code{routes} and values representing elevations}
 
