### 2021-08-30 bcd382a
Remove links in slope_vector() docs for #32
