### 2021-08-30 248d733
add Zurich example
and change urls to https
