+++ NAMESPACE
@@ -14,6 +14,9 @@ export(slope_matrix_weighted)
 export(slope_raster)
 export(slope_vector)
 export(slope_xyz)
+export(z_cumulative_difference)
+export(z_direction)
+export(z_elevation_change_start_end)
 export(z_end)
 export(z_max)
 export(z_mean)
