### 2021-09-03 b274140
Add directed arg to slope_funs for #20
