### 2021-09-03 846cd86
Add directed, progress on #20
