### 2021-09-03 cb0e2e5
Improve docs for slope_xyz()
This fixes another issue in #20 by mentioning alternative functions that can be passed to the fun argument
