### 2021-09-03 e854cb4
Use library(sf) in examples for #37
