+++ R/plot_slope.R
@@ -75,8 +75,9 @@ plot_slope = function(
 #' @param z Elevations at points across a linestring
 #' @inherit plot_slope
 #' @examples
+#' library(sf)
 #' route_xyz = lisbon_road_segment_3d
-#' m = sf::st_coordinates(route_xyz)
+#' m = st_coordinates(route_xyz)
 #' d = cumsum(sequential_dist(m, lonlat = FALSE))
 #' d = c(0, d)
 #' z = m[, 3]
