+++ man/elevation_add.Rd
@@ -38,9 +38,10 @@ library(sf)
 routes = lisbon_road_network[204, ]
 dem = dem_lisbon_raster
 (r3d = elevation_add(routes, dem))
-sf::st_z_range(routes)
-sf::st_z_range(r3d)
-plot(sf::st_coordinates(r3d)[, 3])
+library(sf)
+st_z_range(routes)
+st_z_range(r3d)
+plot(st_coordinates(r3d)[, 3])
 plot_slope(r3d)
 \donttest{
 # Get elevation data (requires internet connection and API key):
