+++ man/slope_raster.Rd
@@ -63,5 +63,7 @@ dem = dem_lisbon_raster
 (s = slope_raster(routes, dem))
 cor(routes$Avg_Slope, s)
 slope_raster(routes, dem, directed = TRUE)
-slope_raster(sf::st_reverse(routes), dem, directed = TRUE)
+# Demonstrate that reverse routes have the opposite directed slope
+library(sf)
+slope_raster(st_reverse(routes), dem, directed = TRUE)
 }
