+++ man/z_value.Rd
@@ -43,7 +43,7 @@ in \code{sfc} objects with \code{XYZ} geometries.
 }
 \examples{
 x = slopes::lisbon_route_3d
-sf::st_geometry(x)
+x
 z_value(x)[1:5]
 xy = slopes::lisbon_route
 try(z_value(xy)) # error message
