### 2021-09-03 f21d4a9
Update dem_raster example for #32
