### 2021-09-03 4c7e207
Update slope_raster example for #37
