### 2021-09-03 fbe5740
Dedupe contents of z_ functions for #37
