### 2021-09-03 4476a38
Reproject result of elevation_get() for #37
