### 2021-09-03 b153e6a
Updates to elevation_extract for #37
