### 2021-09-03 6ba560d
Reduce code duplication in elevation_add() for #37
