### 2021-09-03 a3bfcec
Update to cater for/check sfc objects
