+++ man/elevation_add.Rd
@@ -13,7 +13,7 @@ elevation_add(
 }
 \arguments{
 \item{routes}{Routes, the gradients of which are to be calculated.
-The object must be of class \code{sf} with \code{LINESTRING} geometries.}
+The object must be of class \code{sf} or \code{sfc} with \code{LINESTRING} geometries.}
 
 \item{dem}{Raster overlapping with \code{routes} and values representing elevations}
 
