### 2021-09-03 #420 @Robinlovelace
Exciting update, 48 down, 3 to go! The last three will take some time but are definitely doable, almost there, thanks for you patience. This has been harder but more rewarding than responding to academic peer review comments for @temospena but we're glad for all the feedback and we're almost there. If anyone has any suggestions/support on the final issues/discussion, see our open mega issue here for the unfinished points yet to be responded to: https://github.com/ITSLeeds/slopes/issues/32
