### 2021-09-09 20e8e90
Remove superfluous assignment in lapply loop
