### 2021-09-10 c98c287
Add 2 more tests for slope_distance funs
