### 2021-09-14 ITSLeeds/slopes#32 @Robinlovelace
> ## Package Review
>
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
>     * **Briefly describe any working relationship you have (had) with the
>       package authors.**
>       I shared an office with Robin Lovelace some years ago and we're still in touch.
>
>     * ☒ As the reviewer I confirm that there are no [conflicts of
>       interest](https://devguide.ropensci.org/policies.html#coi) for me to
>       review this work (If you are unsure whether you are in conflict,
>       please speak to your editor _before_ starting your review).
>
>
> #### Documentation
>
> The package includes all the following forms of documentation:
>
>     * ☐ **A statement of need** clearly stating problems the software is
>       designed to solve and its target audience in README
>
>     * ☒ **Installation instructions:** for the development version of
>       package and any non-standard dependencies in README
>
>     * ☒ **Vignette(s)** demonstrating major functionality that runs
>       successfully locally
>
>     * ☒ **Function Documentation:** for all exported functions
>
>     * [half an x] **Examples** (that run successfully locally) for all
>       exported functions
>
>     * ☐ **Community guidelines** including contribution guidelines in the
>       README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and
>       `Maintainer` (which may be autogenerated via `Authors@R`).
>
>
> #### Functionality
>
>     * ☒ **Installation:** Installation succeeds as documented.
>
>     * ☒ **Functionality:** Any functional claims of the software been
>       confirmed.
>
>     * ☒ **Performance:** Any performance claims of the software been
>       confirmed.
>
>     * ☒ **Automated tests:** Unit tests cover essential functions of the
>       package and a reasonable range of inputs and conditions. All tests
>       pass on the local machine.
>
>     * ☐ **Packaging guidelines**: The package conforms to the rOpenSci
>       packaging guidelines
>
>
> Estimated hours spent reviewing: 6-8
>
>     * ☒ Should the author(s) deem it appropriate, I agree to be
>       acknowledged as a package reviewer (“rev” role) in the package
>       DESCRIPTION file.
>
>
> ### Review Comments
>
> Thanks for the invite to peer review. This is a generally awesome
> package that works well and is professionally assembled. I’m a
> reasonably competent R programmer but not an expert in package
> construction or testing, so apologies for any issues with my review.
>
> Tested on a windows machine - runs well.
> #### Comments on the above tickboxes first:
>
>     * Statement of need: does state the problem it solves but not the
>       target audience (the ‘getting started’ article at
>       https://itsleeds.github.io/slopes/articles/slopes.html states the
>       need in a little more detail; the intro to that could go at the top
>       of the readme?)
>
>     * Couldn’t find community guidelines though URL / bug reports are in
>       DESCRIPTION
>
>     * Performance: benchmarks are slower for me but my machine’s quite old
>
>     * Automated tests are present though I don’t feel qualified to confirm
>       they all pass
>
>     * Conforming to rOpenSci packaging guidelines: no CodeMeta file;
>       function and argument naming is all great; package API excellent;
>       code style is compact but great; readme is great, though it does use
>       an example that requires a mapbox API auth key and this isn’t
>       mentioned - more on that below.
>
>
> #### Other comments
>
> Only a couple of comments:
>
>     * The package README includes an example that automatically downloads
>       raster tiles from Mapbox (a fantastically useful feature by the
>       way).
>       (`{r eval = F} lisbon_route_3d_auto = slope_3d(r = lisbon_route)`).
>       However, it doesn’t say that you need a Mapbox API key for it to
>       work. Might be worth adding that, or just a link? (Possibly with a
>       `{r eval = F} usethis::edit_r_environ()` line to show how to add the
>       key?) It’s such a super-useful part of the package, I think it’s
>       worth making sure people know how to use it.
>
>     * The benchmarking test in the README downloads a dem_lisbon.tif file
>       for one of the tests. That file wouldn’t work in either terra or
>       raster for me - and anyway, if we’re comparing the two methods, I’m
>       not sure why it can’t just use the already-available
>       dem_lisbon_raster twice? (That’s how I ran the benchmark.)
>
>
> There are a couple of things I feel might be missing, possibly from the
> README intro or maybe something for the
> https://itsleeds.github.io/slopes/index.html page:
>
>     * Perhaps a little example of how to make your own route to use in the
>       package. It’s quick enough that it wouldn’t bloat the README intro.
>       Or could go in main documentation? E.g. using stplanr:
>
>
> ```
> library(stplanr)
> library(osrm)
>
> #Trip between two Sheffield parks
> #Coords nabbed from right-click in google maps
> trip <- route(
>   from = c(-1.493284835376672,53.39216628908168),
>   to = c(-1.5066026758341629,53.368424563832754),
>   route_fun = osrmRoute,
>   returnclass = "sf"
> )
> #> Most common output is sf
> mapview::mapview(trip)
>
> #Is linestring so can be used in slopes
> class(trip$geometry[1])
>
> #Find slope of route... (using mapbox API)
> sheffield_route_3d_auto = slopes::slope_3d(r = trip)
> slopes::plot_slope(sheffield_route_3d_auto)
> ```
>
>     * It’s a slopes package, not an elevation package, but I wonder if it
>       might be worth mentioning how to find elevation changes? Or provide
>       a simple way to see elevation change summaries for particular
>       slopes? I can imagine that being useful for e.g. assessing hilliness
>       of bike journeys. Excuse my ignorance if this is already in the
>       package - I couldn’t get the distance elevation profile example from
>       ‘getting started’ to work for my Sheffield route above (though in
>       theory I should be able to as sf::st_coordinates gives me the 3D
>       matrix?) I could find some values like this:
>
>
> ```
> library(tidyverse)
> n = sf::st_coordinates(sheffield_route_3d_auto) %>%
>   data.frame %>%
>   mutate(elevationchange = Z-lag(Z))#change in elevation between line segments
>
> #Total net elevation of trip
> sum(n$elevationchange, na.rm = T)
> #Total climb of trip
> sum(n$elevationchange[n$elevationchange > 0], na.rm = T)
> ```
>
> Would it be possible to provide a similar example that ties to the
> distance elevation profile example? Again, sorry - I did try using slope
> package functions to achieve this and failed.
