### 2021-09-14 7c2ac57
Add need for the package and links to READE and intro vignette
