### 2021-09-14 fa00e25
Update vignette with no download data for #32
