### 2021-09-14 a787c7e
Add NEWS.md
