### 2021-09-14 #420 @temospena
Hi!
We just finished reviewing the package.
Please see the responses to reviewers at: https://github.com/ITSLeeds/slopes/issues/32

It had major improvements and we consider that it is much better now.

We started by responding to comments from @ateucher, so many of the comments from @DanOlner were already covered by the previous reply. Any further comments or suggestions are very welcome.

We can write the responses here, if that would be useful.

Thank you for the great suggestions!
