### 2021-09-14 #420 @Robinlovelace
Adding to what Rosa says, thanks for an in depth review guys! There are definitely things that could be changed for the better, we have talked about sub-sampling from vertices to calculate overall segment slope, reducing variability associated with linear segments that run perpendicular to steep slopes, something that @fiftysevendegreesofrad has explored and has an method to handle I think, and we could implement a feature to get only elevation points where they are needed instead of for the whole raster dataset surrounding a particular area, as per https://github.com/ITSLeeds/slopes/issues/21, for example. But overall my feeling is that the package solves the problem that it was set out to solve and I look forward to using it, and see it used in applied settings. We weren't sure what the convention is but if it's to post complete replies to the comments on this issue happy to do that also. Hope you find our changes move the package forward, any follow-ups v. welcome, and looking forward to seeing what new slope datasets generated by this package can do!
