### 2021-09-15 #420 @annakrystalli
Great news!! Congratulations on all the work you've put into this so far! 👏🚀

In terms of your question regarding, it is indeed convention to post replies to each of the authors comments. If some are duplicate, it's fine to respond to them in one go. We do need to make it as easy as possible for reviewers to understand what has been done in response to their comments and, indeed if some comments where not addressed why.  I know it's more effort but it will make it much easier for reviewers to sign it off and it also forms a record of all actions in the review issue itself. Almost there!
