### 2021-09-15 #420 @temospena
- [x] open a parallel branch for revs
- [x] update readme: include more on the description? the main functions?

The README has be substantially updated and now documents the package's main functions: https://github.com/ITSLeeds/slopes/commit/dbacab8792dfc178c37b3ceb5aafd9791f515061

- [x] pass benchmark to a vignete
- [x] Update argument names
- [x] Don't export plot_dz

More detailed responses in-line

# Response to review 2:

> ## Package Review
>
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
>     * **Briefly describe any working relationship you have (had) with the package authors.**
>
>     * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
>
> #### Documentation
>
> The package includes all the following forms of documentation:
>
>     * [x]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
>
>     * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
>
>     * [ ]  **Vignette(s)** demonstrating major functionality that runs successfully locally
>
>     * [x]  **Function Documentation:** for all exported functions
>
>     * [ ]  **Examples** (that run successfully locally) for all exported functions
>
>     * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
>
> #### Functionality
>
>     * [x]  **Installation:** Installation succeeds as documented.
>
>     * [x]  **Functionality:** Any functional claims of the software been confirmed.
>
>     * [x]  **Performance:** Any performance claims of the software been confirmed.
>
>     * [x]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
>     * [ ]  **Packaging guidelines:**: The package conforms to the rOpenSci packaging guidelines
>
>
> Estimated hours spent reviewing: 8-10
>
>     * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
>
>
> ## Review Comments
>
> This is a great package that fills an obvious need. It has all of the necessary functionality required to do the job it sets out to do - calculate slopes of routes. The functions for getting and using the appropriate DEM for an area removes a huge barrier for users, so kudos! I had fun test-driving the package and learned a lot. I hope my review doesn't come across as negative - it really is an excellent package and with some rough edges filed down I think it could be even better! I did get a bit in the weeds at times - so while there are a lot of comments most of them are quite small!
> ## Review Checklist
> ### Documentation
> #### Statement of Need
>
> The introductory sentence in the README gives a very brief description of what the package does, and the target audience is fairly easily inferred (i.e., people who need to calculate slopes of linear features). I don't think this needs much more as the application is fairly clear, however you could include an overview of some of the other things the package does to help a user get to the point of calculating slopes (e.g., getting elevation data from online sources, extracting elevation points), as well as the visualization tools (plotting functions). I like that it compares the package's functionality to other common tools (ESRI's `Slope3d()`).

>
>     * Installation instructions work and are straightforward, however it would be helpful to add some information about MapBox, and that an API key is needed for some functions. Also note that for some functions, there are packages that will need to be installed separately as they are not installed automatically (i.e., are in `Suggests` - `ceramic`, `ggdist`, etc.)

- [x] Vignette(s)

>
>     * The main vignette (`slopes.Rmd`) doesn't actually show any examples of what the stated main purpose of the package is - to provide slopes of linear features from a DEM. The README has the most comprehensive walkthrough of the main advertised functionality of the package. It is available on GitHub and the pkgdown page - but not available in R as a vignette. I suggest taking the README content and creating a vignette (likely the primary one), or perhaps even better - add it to the current `slopes.Rmd` so you combine the theory with a demonstration of the implementation.
>
>     * `verification.Rmd`:
>
>       * a `library(raster)` call is needed to plot the raster object (`plot(e)` on [line 91](https://github.com/ITSLeeds/slopes/blob/3a25f7ce11ced3be313c0eadbda395551887fa2e/vignettes/verification.Rmd#L91))
>       * `slope_raster(r, e = e)` on [line 123](https://github.com/ITSLeeds/slopes/blob/3a25f7ce11ced3be313c0eadbda395551887fa2e/vignettes/verification.Rmd#L123) fails because `points2line_trajectory()` does not assign CRS to the output - as a result, `lonlat` is not determined internally (more on this below). I made it work here by tweaking `points2line_trajectory()`:
>
>       ```
>       points2line_trajectory = function(p) {
>         c = st_coordinates(p)
>         i = seq(nrow(p) - 2)
>         l = purrr::map(i, ~ sf::st_linestring(c[.x:(.x + 1), ]))
>         lfc = sf::st_sfc(l)
>         a = seq(length(lfc)) + 1 # sequence to subset
>         p_data = cbind(sf::st_set_geometry(p[a, ], NULL))
>         sf::st_sf(p_data, geometry = lfc, crs = st_crs(p)) # set the crs here
>       }
>       ```
>

We have made substantial improvements to the vignettes. There is now an updated introductory `slopes.Rmd` vignette that demonstrates all key function and a shorter README that links to the other vignettes. The more theoretical introduction to slopes has been moved into a new vignette, `intro-to-slopes`. See here for details: https://github.com/ITSLeeds/slopes/pull/38

## Function Documentation

>
> **General Comments:**

- [x] Package-level documentation.

>
>     * Package-level documentation would be convenient, so a user can type `?slopes` and get an overview of what the package does, authors, urls etc. - and link to the package index). You can set this up with `usethis::use_package_doc()`.

This has been done in https://github.com/ITSLeeds/slopes/commit/018d2d60d0d8065d9d5865c98752b34cd14806bf. Many thanks for the tip.

- [x] `@return` sections
>
>     * All function documentation should have a `@value` (return) section so the user knows what they can expect from the function.

- [x] `\value{...}` sections have been added to the .Rd files using `#' @return` in the source code thanks to `roxygen2`. See commit https://github.com/ITSLeeds/slopes/commit/7a8cdc519828aae10ca7d4d4bf3e20a2ca11683f

- [x] Datasets

>
>     * The included datasets are great - nice tidy real-world examples to play with.

- [x] Dataset docs

>
>     * The datasets could be documented a bit more clearly - how many road segments, lengths/distances etc. It's nice to have the context when you are playing with the examples. It is not obvious to the user the differences between the various datasets with Lisbon road segments.

This is a good point, we put the dataset docs together in a rush. They have been iteratively improved, including in this commit: 45068459072a8602985e22e5d7730d24122a1cdd

- [x] Argument names

>
>     * In a lot of cases the names of function arguments could be clearer - many are single letters that don't have an obvious meaning.

Thanks for the comment. We agree that the argument names were at times overly terse and we now find ourselves wishing we spent more time thinking about this earlier on. The good news is that, after some initial issues, we have now got updated and consolidated argument names, as documented here: https://github.com/ITSLeeds/slopes/pull/31/files. To summarise that we now have:

- `@param routes Routes, the gradients of which are to be calculated.` replacing `r`
- `@param dem Raster overlapping with `routes` and values representing elevations` replacing `e`
- `@param elevations Elevations in same units as x (assumed to be metres)` also replacing `e`

We also updated arg names in the plotting functions in https://github.com/ITSLeeds/slopes/commit/7651276edab847dd0ce27d33ed182669ae3b28a9

- [x] Updated description: 311af41fcced0157ab7e311008fef8b2570d20c8

>     * Data documentation should provide clearer sources for the data - for example, `dem_lisbon_raster` has only a link to a `terra` GitHub issue. The various Lisbon road segment datasets link to the ESRI's 3D Analyst extension webpage - which is not the source of the data, rather the tool that created it.

- [x] Documentation for `elevation_get()`

>
> **Specific documentation issues:**
>
>     * `elevations_get()` could use more explanation of how these elevations are obtained and how to customize it. What web service(s) is it using? I see that it uses MapBox via `ceramic::cc_elevation()` however that might be a bit opaque to a user. I think at least a mention that they need a MapBox API key (ideally with a link/instructions on how to get one), and a link to the `ceramic::cc_elevation()` documentation would be helpful, as well as some examples showing customization by passing options to `cc_elevation()`.

The documentation of this function has been improved, as can be seen in https://itsleeds.github.io/slopes/reference/elevation_get.html. Specifically, we link to long form documentation explaining how to get an API key, rather than duplicating content in the function documentation. We also link explicitly to the `cc_elevation()` page for people who want more details. Note: we do not see this functionality, that is reliant on potentially temperamental web APIs, as core to the package so we do not want to emphasise this aspect of the package, believing that it's better if users get and understand their own DEM data using other tools, in line with the Linux philosophy of modularity. See https://github.com/ITSLeeds/slopes/commit/7a67e2db3caeccf42c693067076ce224cbaf2de1 for details on updated links to the ceramic package.

- [x] `elevation_extract()` documentation

>
>     * `elevation_extract()` documentation would benefit from more explanation of its arguments: what other options can be used for `method`? What are the benefits of using `terra` (and what is the alternative? - I know it's `raster` but the user may not).

The documentation has been improved as a result of this comment. See https://github.com/ITSLeeds/slopes/commit/18b0a5dbd7626b0b05029bfcd298bde38fd16796 for details.

One question for the reviewers: should this function even be exported? It seems of little additional use compared with the original `extract()` functions for users but we thought it worth asking and minimising changes. If nothing else the documentation is now more interesting!

- [x] `plot_dz`

>
>     * `plot_dz`:
>
>       * All argument documentation should be expanded, at a minimum stating the defaults, the accepted options (and/or type/format of inputs).

- [x] Agreed. To ease maintenance, `@param` tags have been shifted up into the exported function `plot_slope()`. All parameters are now better documented, with improved names (`legend_position` instead of `x`, for example) and default values.

>       * I tend to think that the [setting of the default colour](https://github.com/ITSLeeds/slopes/blob/9365709a95312b4feff3689a9e219c06af314167/R/plot_slope.R#L63-L66) palette (`p`) could be done internally in the function, and set the default argument value to `NULL` - this would be a simpler interface for the user. In addition, the value of `requireNamespace()` does not need to be assigned to anything in this `ifelse()` statement.

- [x] The interface has been simplified and we removed the `ifelse` statement. The parameter is still set with a default but now that `colorspace` is imported it is guaranteed to work.

>       * It is not clear (to me at least) what the argument `x` does - I can infer that it is legend position, but neither the name of the argument nor the description tell you this. I suggest renaming it and explaining what it does.

- [x] `plot_slope()` arguments

>       * It's not clear what the values passed to `s` do, nor what significance the default values `3:18` hold (especially as they relate to `brks`).

The parameter name has been clarified and renamed:

```
#' @param seq_brks Sequence of breaks to show in legend.
#'   Includes negative numbers and omits zero by default
```

- [x] `plot_slope()` changes

>
>     * `plot_slope()` - Similar comments on argument names, defaults and descriptions as `plot_dz`. It seems to me that this is the function that users would call much more often than manually constructing the components and then using `plot_dz()` - if that's the case, would you consider not exporting `plot_dz`, or directing users to `plot_slope()` in most cases?

Agreed, the `plot_slope()` function is the one that people will use and it was confusing having both `plot_slope()` and the more esoteric helper function `plot_dz` exported. Only `plot_slope()` is now exported and the documentation is better. Also, the colourscheme has been updated so 0 corresponds with white, which was not previously the case.

- [x] `sequential_dist()`

>     * `sequential_dist()` - Some more clarity on what `m` should look like would be helpful. What are the dimensions of the matrix, what are the order of the columns? The example doesn't give much clarity on this.

The description of the input `m` has been updated and now provides this important information:

```r
#' @param m Matrix containing coordinates and elevations.
#'   The matrix should have three columns: x, y, and z. Typically
#'   these correspond to location in the West-East, South-North, and vertical
#'   elevation axes respectively.
#'   In data with geographic coordinates, Z values are assumed to be in
#'   metres. In data with projected coordinates, Z values are assumed to have
#'   the same units as the X and Y coordinates.
```

Note: this change also benefits the `slope_matrix()` functions. See https://github.com/ITSLeeds/slopes/commit/7637f45fe1d9f2d18d3cfdd7362f000ce3c5edb6 for details.

- [x] (now `elevation_add()`) - this is now described in README and in Get Started vignette. d9c40cac95ca8e106ed578daf85edc1a4f450fe4

>     * `slope_3d()` - it is not clear from the documentation that if you omit the raster object (`e = NULL`) then the function will download the necessary raster from MapBox - this is amazing functionality and should be advertised!

- [x] `?slope_vector`

>     * The documentation links in `?slope_vector` are not necessary since they link to the same help page.
>

Agreed, this has been fixed in https://github.com/ITSLeeds/slopes/commit/bcd382adaa8bd230b47a6ca8bce9286f937e961d

- [x] `slope_matrix_mean()`

>     * I anticipated that there would be a `slope_matrix_mean()` that calculates the unweighted mean slope from a matrix, in addition to `slope_matrix()` and `slope_matrix_weighted()` - analogous to `slope_distance_mean()`.

This function has been added: https://itsleeds.github.io/slopes/reference/slope_matrix.html

- [x] `slope_xyz`

>     * `slope_xyz` - Alternative functions that can be passed to `fun` should be documented.

Good point. This, plus other improvement to the docs for this function, have been made in https://github.com/ITSLeeds/slopes/commit/cb0e2e56127f6dedd83744e8b515581b373d4abd - wondering if the function should ever have been exported but it's there now and think it could be useful for others. Especially now it's better documented!

> #### Examples
>
> **General Comments:**
>
>     * All functions have examples that show the basic usage.

- [x] How functions are called
>
>     * I think it would be good to standardize how functions are called in examples (sometimes they are called with `package::function()` and sometimes just `function()`). I think for `slopes` functions you probably don't need to use the `::` notation, as it assumed the package is loaded, and for functions from other packages you should standardize on either using (for example) `library(sf)` or `sf::function()` (I would lean towards the former to avoid the method dispatch issues I flagged in several examples below).

We call `sf` and tend to avoid the `::` in the examples, although believe it is useful in some places in the package, e.g. to concisely to demonstrate that a function is from a particular package. See  https://github.com/ITSLeeds/slopes/commit/e854cb4c7d6e0d701e7170d4b05749bef3a137b4 for details.

- [x] Explanation of examples

>
>     * In general, I think the examples could contain a bit more explanation of what is happening and/or use more meaningful values in the examples (e.g., real lon/lat pairs and distances rather than simple integers) - this would facilitate understanding the intention and outputs.
>

We think that the use of simple interger values in the examples supports understanding by allowing people to see how they work from first principles, without needing to think of long lon/lat values. The examples in the vignettes, and the examples in the functions that do the 'heavy lifting' that are most commonly used by users, have real world examples. We would be happy to improve any specific examples that are not clearly explained. Overall, we have improved the documentation but we would welcome any further guidance on examples that should be improved.

- [x] Examples that do not run

>     * Rather than commenting out examples that should not be automtically run, it would be better to wrap them in `\dontrun{}` or `\donttest{}`.
>

This is a good point that would also cause issues for the CRAN submission. Addressed in https://github.com/ITSLeeds/slopes/commit/fe5f12da7250f99b08921bd7cf90928a497242c2

- [x] `route_cyclestreets` docs

>
> **Specific Issues:**
>
>     * In the example for `cyclestreets_route`, I believe this contains the (commented out) code for generating the data, not an example of its use. I think this could be confusing for the user, and should probably belong in a script in the `data-raw` folder - the code also uses the `stplanr` package, which is not listed as a dependency in `slopes`. In order to run the commented-out `plot()` call, the `sf` package must be loaded to use the correct `plot()` method.

This has been fixed, as can be seen here: https://itsleeds.github.io/slopes/reference/cyclestreets_route.html

- [x] `dem_lisbon_raster` example

>
>     * In the `dem_lisbon_raster()` example the `sf` package must be loaded to use the correct `plot()` method.
>

- [x] `slope_raster()` example

>     * In the example for `slope_raster()`, the `sf` package must be loaded for the correct `[` method to be used (otherwise `r = lisbon_road_segments[1:3, ]` results in an invalid `sf` object).
>

This has been fixed in https://github.com/ITSLeeds/slopes/commit/4c7e2077d8fe271ed08804e66b6daecb606311ab

- [x] CONTRIBUTING and CODE OF CONDUCT added,

>
> #### Community guidelines
>
>     * DESCRIPTION file has URLs for GitHub and pkgdown pages, and a BugReports field +1
>
>     * There is currently no CONTRIBUTING file or instructions on how to contribute in the README.
>
>     * Currently there is no Code of Conduct, though it appears that will be [added upon acceptance to rOpenSci](https://devguide.ropensci.org/collaboration.html?q=conduct#coc-file).

- [x] In progress

>
>
> ### Functionality
>
[…]
