### 2021-09-20 #420 @ateucher
My congratulations to @temospena and @Robinlovelace - that was a ton of work, and I think it's much better. I hope you found it worth it! In my opinion you've more than adequately addressed my comments and **_I'm happy to recommend acceptance_**. I especially love the new suite of vignettes, in particular `roadnetworkcycling.Rmd`.  In the cases where you chose not to implement my suggestions, your reasons were well explained and I'm satisfied with that.

As to the question of whether to export `elevation_extract()` - I don't see any harm in exporting it even if it is a thin wrapper, especially now that you've documented nicely.

One element of my review that I didn't see an answer to was `"How does one know/set the direction a route is meant to be traveled (or for a river, the direction of flow)?"` - however I think now that this is out of scope for this package and is maybe answered elsewhere (perhaps `stplanr`)?

Thanks again for the opportunity to review, and kudos for a great package!

Andy
