### 2021-09-27 #420 @ropensci-review-bot
Approved! Thanks @temospena for submitting and @DanOlner, @ateucher for your reviews! :grin:
