### 2021-09-27 #420 @Robinlovelace
This is fantastic news, many thanks @annakrystalli and all! One quick question: what is the process for submission to JOSS at this point? Discussing next steps with Rosa now!
