### 2021-09-27 #420 @Robinlovelace
Heads-up, just put in a PR that changes the URLs in the docs: https://github.com/ITSLeeds/slopes/pull/42

Please take a look Anna and let us know if that looks good.
