### 2021-09-27 #420 @Robinlovelace
Exciting news, it's on rOpenSci :tada: https://github.com/ropensci/slopes
