### 2021-09-27 #420 @temospena
To keep tracking:

 * If you already had a `pkgdown` website **and are ok relying only on [rOpenSci central docs building and branding](https://devguide.ropensci.org/ci.html#even-more-ci-ropensci-docs)**,

- [x]   deactivate the automatic deployment you might have set up
- [ ]   remove styling tweaks from your pkgdown config but keep that config file
- [x]   replace the whole current `pkgdown` website with a [redirecting page](https://devguide.ropensci.org/redirect.html)
- [x]   replace your package docs URL with `https://docs.ropensci.org/package_name`
- [x]   In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`

 * [ ]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`. If Appveyor does not pick up new commits after transfer, you might need to delete and re-create the Appveyor project. (Repo transfers are smoother with GitHub Actions)
 * [ ]  Please check you have updated the package version to a post-review version and that you documented all changes in NEWS.md
 * [ ]  We're starting to roll out software metadata files to all rOpenSci packages via the Codemeta initiative, see https://docs.ropensci.org/codemetar/for how to include it in your package, after installing the package - should be easy as running `codemetar::write_codemeta()` in the root of your package.
