### 2021-09-27 #420 @Robinlovelace
Looking at the outstanding tasks...

- [ ] remove styling tweaks from your pkgdown config but keep that config file

Think that's done, not sure we ever had any styling tweaks.

- [x] Replace the whole current pkgdown website with a redirecting page

Not done but not in any rush to do this as more important that https://github.com/itsleeds/slopes redirects at this stage I think, seems to be the top hit on Google.

- [x] Fix any links in badges for CI and coverage to point to the ropensci URL.

Done

- [x] Please check you have updated the package version to a post-review version and that you documented all changes in NEWS.md
- [x] We're starting to roll out software metadata files to all rOpenSci packages via the Codemeta initiative, see https://docs.ropensci.org/codemetar/for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package.
