### 2021-09-27 #420 @Robinlovelace
One comment that was not addressed by @ateucher was whether the package supports directed slopes. Answer, yes: it does now! See here for info (I've just fixed the docs but they are taking time to update on rOpenSci): https://docs.ropensci.org/slopes/articles/slopes.html#splitting-the-network-1
