### 2021-10-18 #420 @fiftysevendegreesofrad
Hi all,

this package looks great!

Further to @Robinlovelace 's comment above, I have indeed been working on ways to auto-correct heights where interpolation of steep terrain results in lots of spurious elevation change. This rather seriously affected our cycling models of Chepstow where plenty of routes go along the top or bottom of cliffs; in that particular case I could correct them by hand but I'm pleased to say I can now offer a general solution.

Temporary link to the draft paper here - please don't share more widely as it needs to go through peer review https://www.dropbox.com/s/b7id0n3g594po18/bayesian%20drape%20draft.pdf?dl=0 and repo here https://github.com/fiftysevendegreesofrad/BayesianDrape

Comments are welcome - for one thing I should add a citation to the slopes R package, please let me know your preferred citation :)
